@@ -10,7 +10,7 @@
 
 namespace SmartFactory;
 
-use SmartFactory\Interfaces\ILanguageManager;
+use \SmartFactory\Interfaces\ILanguageManager;
 
 /**
  * Class for working with localization of texts.
@@ -26,8 +26,27 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    protected $localization_path;
+    protected string $localization_path = "";
+
+    /**
+     * Internal variable for storing the paths of the additional localization files.
+     *
+     * @var string|array
+     *
+     * @author Oleg Schildt
+     */
+    protected string|array $additional_localization_files = [];
     
+    /**
+     * Internal variable for storing the fallback language.
+     * If set and a translation is missing on a language, the translation on this language will be used.
+     *
+     * @var string
+     *
+     * @author Oleg Schildt
+     */
+     protected string $use_fallback_language = "";
+
     /**
      * Internal variable for storing the current context.
      *
@@ -37,35 +56,53 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    static protected $context = "default";
-    
+    static protected string $context = "default";
+
     /**
-     * Internal variable for storing the current language.
+     * Internal variable for storing the state whether the dictionary is loaded or not.
+     *
+     * @var bool
+     *
+     * @author Oleg Schildt
+     */
+    protected bool $dictionary_loaded = false;
+
+    /**
+     * Internal variable for storing the state whether the APCU should be used.
+     *
+     * @var bool
+     *
+     * @author Oleg Schildt
+     */
+    protected bool $use_apcu = false;
+
+    /**
+     * Internal variable for storing the state whether the last selected language can be stored to cookie.
+     *
+     * @var bool
+     *
+     * @author Oleg Schildt
+     */
+    protected bool $use_cookie = false;
+
+    /**
+     * Internal variable for storing the cookie path.
      *
      * @var string
      *
      * @author Oleg Schildt
      */
-    static protected $current_language = "";
-    
+    protected string $cookie_path = "/";
+
     /**
-     * Internal variable for storing the state whether the dictionary is loaded or not.
+     * Internal variable for storing the state whether the E_USER_NOTICE is triggered in the case of missing translations.
      *
-     * @var boolean
+     * @var bool
      *
      * @author Oleg Schildt
      */
-    static protected $dictionary_loaded = false;
-    
-    /**
-     * Internal variable for storing the state whether the APCU should be used.
-     *
-     * @var boolean
-     *
-     * @author Oleg Schildt
-     */
-    protected $use_apcu = false;
-    
+    protected bool $warn_missing = true;
+
     /**
      * Internal array for storing the list of supported languages.
      *
@@ -73,8 +110,17 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    static protected $supported_languages = [];
-    
+    static protected array $supported_languages = [];
+
+    /**
+     * Internal variable for storing the current language.
+     *
+     * @var array
+     *
+     * @author Oleg Schildt
+     */
+    static protected array $current_language = [];
+
     /**
      * Internal array for storing the list of language name translations.
      *
@@ -82,8 +128,8 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    static protected $languages = [];
-    
+    static protected array $languages = [];
+
     /**
      * Internal array for storing the list of country name translations.
      *
@@ -91,8 +137,8 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    static protected $countries = [];
-    
+    static protected array $countries = [];
+
     /**
      * Internal array for storing the list of text translations.
      *
@@ -100,8 +146,8 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    static protected $texts = [];
-    
+    static protected array $texts = [];
+
     /**
      * Initializes the language manager with parameters.
      *
@@ -110,31 +156,72 @@ class LanguageManager implements ILanguageManager
      *
      * - $parameters["localization_path"] - the path where the localization files are stored.
      *
+     * - $parameters["use_cookie"] - if set to true, the last selected language is stored to cookie.
+     *
+     * - $parameters["use_fallback_language"] - if set and a translation is missing on a language, the translation on this language will be used.
+     *
+     * - $parameters["cookie_path"] - Cookie path.
+     *
      * - $parameters["use_apcu"] - if installed, apcu can be used to cache the translations in the memory.
      *
-     * @return boolean
-     * Returns true upon successful initialization, otherwise false.
+     * - $parameters["warn_missing"] - If it is set to true, the E_USER_NOTICE is triggered in the case of missing translations.
+     *
+     * @return void
+     *
+     * @throws \Exception
+     * It might throw an exception in the case of any errors.
      *
      * @author Oleg Schildt
      */
-    public function init($parameters)
+    public function init(array $parameters): void
     {
         if (!empty($parameters["localization_path"])) {
             $this->localization_path = $parameters["localization_path"];
         }
-        
+
         if (!empty($parameters["use_apcu"])) {
             $this->use_apcu = $parameters["use_apcu"];
         }
-        
-        return true;
+
+        if (!empty($parameters["use_cookie"])) {
+            $this->use_cookie = $parameters["use_cookie"];
+        }
+
+        if (!empty($parameters["cookie_path"])) {
+            $this->cookie_path = $parameters["cookie_path"];
+        }
+
+        if (!empty($parameters["use_fallback_language"])) {
+            $this->use_fallback_language = $parameters["use_fallback_language"];
+        }
+
+        if (!empty($parameters["warn_missing"])) {
+            $this->warn_missing = $parameters["warn_missing"];
+        }
     }
-    
+
     /**
-     * This is internal auxiliary function for loading the translations from the source JSON file.
+     * Adds additional localization files to the dictionary.
      *
-     * @return boolean
-     * It should return true if the dictoinary has been successfully loaded, otherwise false.
+     * @param string $localization_file
+     * The path of the additional localization file.
+     *
+     * @return void
+     *
+     * @throws \Exception
+     * It might throw an exception in the case of any errors.
+     *
+     * @author Oleg Schildt
+     */
+    public function addLocalizationFile(string $localization_file): void
+    {
+        $this->additional_localization_files[] = $localization_file;
+    }
+
+    /**
+     * This is function for loading the translations from the source JSON file.
+     *
+     * @return void
      *
      * @throws \Exception
      * It might throw the following exceptions in the case of any errors:
@@ -142,29 +229,24 @@ class LanguageManager implements ILanguageManager
      * - if the translation file is invalid.
      *
      * @author Oleg Schildt
-     *
-     * @since 3.1.1
-     * Important change 3.
-     * @since 3.1.2
-     * Important change 4.
      */
-    public function loadDictionary()
+    public function loadDictionary(): void
     {
-        if (self::$dictionary_loaded) {
-            return true;
+        if ($this->dictionary_loaded) {
+            return;
         }
-        
+
         if ($this->use_apcu) {
             do {
                 if (!apcu_exists("dictionary_supported_languages")) {
                     break;
                 }
-                
+
                 self::$supported_languages = apcu_fetch("dictionary_supported_languages");
                 if (empty(self::$supported_languages)) {
                     break;
                 }
-                
+
                 if (!apcu_exists("dictionary_languages")) {
                     break;
                 }
@@ -172,7 +254,7 @@ class LanguageManager implements ILanguageManager
                 if (empty(self::$languages)) {
                     break;
                 }
-                
+
                 if (!apcu_exists("dictionary_countries")) {
                     break;
                 }
@@ -180,92 +262,162 @@ class LanguageManager implements ILanguageManager
                 if (empty(self::$countries)) {
                     break;
                 }
-                
+
                 if (!apcu_exists("dictionary_texts")) {
                     break;
                 }
+
                 self::$texts = apcu_fetch("dictionary_texts");
                 if (empty(self::$texts)) {
                     break;
                 }
-                
-                return true;
+
+                return;
             } while (false);
         }
-        
+
+        $json_array = [];
+
+        $json = file_get_contents($this->localization_path . "config.json");
+        if ($json === false) {
+            throw new \Exception("Translation file '" . $this->localization_path . "config.json" . "' cannot be loaded or does not exist!");
+        }
+
+        try {
+            json_to_array($json, $json_array);
+        } catch (\Throwable $ex) {
+            throw new \Exception("Translation file '" . $this->localization_path . "config.json" . "' is invalid!" . "\n\n" . $ex->getMessage());
+        }
+
+        $json = file_get_contents($this->localization_path . "languages.json");
+        if ($json === false) {
+            throw new \Exception("Translation file '" . $this->localization_path . "languages.json" . "' cannot be loaded or does not exist!");
+        }
+
+        try {
+            $json_array["languages"] = [];
+            json_to_array($json, $json_array["languages"]);
+        } catch (\Throwable $ex) {
+            throw new \Exception("Translation file '" . $this->localization_path . "languages.json" . "' is invalid!" . "\n\n" . $ex->getMessage());
+        }
+
+        $json = file_get_contents($this->localization_path . "countries.json");
+        if ($json === false) {
+            throw new \Exception("Translation file '" . $this->localization_path . "countries.json" . "' cannot be loaded or does not exist!");
+        }
+
+        try {
+            $json_array["countries"] = [];
+            json_to_array($json, $json_array["countries"]);
+        } catch (\Throwable $ex) {
+            throw new \Exception("Translation file '" . $this->localization_path . "countries.json" . "' is invalid!" . "\n\n" . $ex->getMessage());
+        }
+
         $json = file_get_contents($this->localization_path . "texts.json");
         if ($json === false) {
             throw new \Exception("Translation file '" . $this->localization_path . "texts.json" . "' cannot be loaded or does not exist!");
         }
-        
-        $json_array = [];
+
         try {
-            json_to_array($json, $json_array);
+            $json_array["texts"] = [];
+            json_to_array($json, $json_array["texts"]);
         } catch (\Throwable $ex) {
             throw new \Exception("Translation file '" . $this->localization_path . "texts.json" . "' is invalid!" . "\n\n" . $ex->getMessage());
         }
-        
+
+        if (!empty($this->additional_localization_files)) {
+            foreach ($this->additional_localization_files as $localization_file) {
+                $json = file_get_contents($localization_file);
+                if ($json === false) {
+                    throw new \Exception("Translation file '" . $localization_file . "' cannot be loaded or does not exist!");
+                }
+
+                try {
+                    $translation_texts = [];
+                    json_to_array($json, $translation_texts);
+
+                    $json_array["texts"] = array_merge($json_array["texts"], $translation_texts);
+                } catch (\Throwable $ex) {
+                    throw new \Exception("Translation file '" . $localization_file . "' is invalid!" . "\n\n" . $ex->getMessage());
+                }
+            }
+        }
+
         if (!empty($json_array["interface_languages"])) {
             foreach ($json_array["interface_languages"] as $lang_code) {
                 self::$supported_languages[$lang_code] = $lang_code;
             }
         }
-        
+
         if (!empty($json_array["languages"])) {
-            foreach ($json_array["languages"] as $text_id => &$translations) {
+            foreach ($json_array["languages"] as $text_id =>&$translations) {
                 foreach ($translations as $lang_code => $translation) {
                     self::$languages[$lang_code][$text_id] = $translation;
                 }
             }
         }
-        
+
         if (!empty($json_array["countries"])) {
-            foreach ($json_array["countries"] as $text_id => &$translations) {
+            foreach ($json_array["countries"] as $text_id => $translations) {
                 foreach ($translations as $lang_code => $translation) {
                     self::$countries[$lang_code][$text_id] = $translation;
                 }
             }
         }
-        
+
         if (!empty($json_array["texts"])) {
-            foreach ($json_array["texts"] as $text_id => &$translations) {
+            foreach ($json_array["texts"] as $text_id => $translations) {
                 foreach ($translations as $lang_code => $translation) {
                     self::$texts[$lang_code][$text_id] = $translation;
                 }
             }
         }
-        
-        self::$dictionary_loaded = true;
-        
-        foreach (self::$supported_languages as $lng) {
-            self::$current_language = $lng;
-            break;
-        }
-        
+
+        $this->dictionary_loaded = true;
+
         if ($this->use_apcu) {
             apcu_store("dictionary_supported_languages", self::$supported_languages);
             apcu_store("dictionary_languages", self::$languages);
             apcu_store("dictionary_countries", self::$countries);
             apcu_store("dictionary_texts", self::$texts);
         }
-        
-        return true;
     } // loadDictionary
-    
+
+    /**
+     * Adds additional translations to the dictionary.
+     *
+     * @param array $dictionary
+     * The array with additional translations.
+     *
+     * @return void
+     *
+     * @throws \Exception
+     * It might throw an exception in the case of any errors.
+     *
+     * @author Oleg Schildt
+     */
+    public function extendDictionary(array $dictionary): void
+    {
+        $this->loadDictionary();
+        
+        foreach ($dictionary as $text_id => $translations) {
+            foreach ($translations as $lang_code => $translation) {
+                self::$texts[$lang_code][$text_id] = $translation;
+            }
+        }
+    }
+
     /**
      * This function should detect the current language based on cookies, browser languages etc.
      *
      * Priority:
      *
      * 1. explicitly set by the request parameter language.
-     * 2. last language in the session.
+     * 2. header 'Content-Language'.
      * 3. last language in the cookie.
      * 4. browser default language.
      * 5. the first one from the supported list.
      * 6. English.
-     *
-     * @param string $context
-     * The context of the application.
      *
      * Some applications may consist of two parts - administration
      * console and public site. A usual example is a CMS system.
@@ -284,64 +436,85 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    public function detectLanguage($context = "default")
+    public function detectLanguage(): void
     {
-        self::$context = $context;
-        
-        // Priority:
-        
-        // 1) explicitly set by the request parameter language
-        // 2) last language in the session
-        // 3) last language in the cookie
-        // 4) browser default language
-        // 5) the first one from the supported list
-        // 6) English
-        
         // Let's go
-        
+
+        // 6. English
         $language = "en";
-        
-        // 5) the first one from the supported list
+
+        // 5. the first one from the supported list
         foreach (self::$supported_languages as $lng) {
             $language = $lng;
             break;
         }
-        
-        // 4) browser default
+
+        // 4. browser default language
         if (isset($_SERVER["HTTP_ACCEPT_LANGUAGE"]) && trim($_SERVER["HTTP_ACCEPT_LANGUAGE"]) != "") {
             $accepted = explode(',', $_SERVER["HTTP_ACCEPT_LANGUAGE"]);
-            
-            foreach ($accepted as $key => $name) {
+
+            foreach ($accepted as $name) {
                 $code = explode(';', $name);
                 // handle the cases like en-ca => en
                 $code = explode("-", $code[0]);
-                
+
                 if (!empty(self::$supported_languages[$code[0]])) {
                     $language = $code[0];
                     break;
                 }
             }
         }
-        
-        // 3) last language in the cookie
+
+        // 3. last language in the cookie
         $tmp = get_cookie(self::$context . "_language");
         if (!empty($tmp) && !empty(self::$supported_languages[$tmp])) {
             $language = $tmp;
         }
-        
-        // 2) last language in the session
-        if (!empty(session()->vars()[self::$context . "_language"]) && !empty(self::$supported_languages[session()->vars()[self::$context . "_language"]])) {
-            $language = session()->vars()[self::$context . "_language"];
+
+        // 2. header 'Content-Language'.
+        $header = get_header("Content-Language");
+        if (!empty($header) && !empty(self::$supported_languages[$header])) {
+            $language = $header;
         }
-        
-        // 3) explicitly set by request parameter language
+
+        // 1. explicitly set by request parameter language
         if (!empty($_REQUEST["language"]) && !empty(self::$supported_languages[$_REQUEST["language"]])) {
             $language = $_REQUEST["language"];
         }
-        
+
         $this->setCurrentLanguage($language);
     } // detectLanguage
-    
+
+    /**
+     * Sets the current context.
+     *
+     * Some applications may consist of two parts - administration
+     * console and public site. A usual example is a CMS system.
+     *
+     * For example, you are using administration console in English
+     * and editing the public site for German and French.
+     * When you open the public site for preview in German or French,
+     * you want it to be open in the corresponding language, but
+     * the administration console should remain in English.
+     *
+     * With the help of $context, you are able to maintain different
+     * languages for different parts of your application.
+     * If you do not need the $context, just do not specify it.
+     *
+     * @param string $context
+     * The name of the context.
+     *
+     * @return void
+     *
+     * @see LanguageManager::getContext()
+     *
+     * @author Oleg Schildt
+     */
+    public function setContext(string $context): void
+    {
+        self::$context = $context;
+    } // setContext
+
     /**
      * Returns the current context.
      *
@@ -358,16 +531,18 @@ class LanguageManager implements ILanguageManager
      * languages for different parts of your application.
      * If you do not need the $context, just do not specify it.
      *
-     * @return boolean
+     * @see LanguageManager::setContext()
+     *
+     * @return string
      * Returns the current context.
      *
      * @author Oleg Schildt
      */
-    public function getContext()
+    public function getContext(): string
     {
         return self::$context;
     } // getContext
-    
+
     /**
      * Returns the list of supported languages.
      *
@@ -376,39 +551,50 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    public function getSupportedLanguages()
+    public function getSupportedLanguages(): array
     {
         return self::$supported_languages;
     } // getSupportedLanguages
-    
+
     /**
      * Sets the current language.
      *
      * @param string $language
      * The language ISO code to be set.
      *
-     * @return boolean
-     * Returns true if the current language has been successfully set, otherwise false.
+     * @return void
      *
      * @see LanguageManager::getCurrentLanguage()
      *
      * @author Oleg Schildt
      */
-    public function setCurrentLanguage($language)
+    public function setCurrentLanguage(string $language): void
     {
         if (empty(self::$supported_languages[$language])) {
-            return false;
+            return;
         }
-        
-        self::$current_language = $language;
-        
-        session()->vars()[self::$context . "_language"] = $language;
-        
-        set_cookie(self::$context . "_language", $language, time() + 365 * 24 * 3600,  ["samesite" => "strict"]);
-        
-        return true;
+
+        self::$current_language[self::$context] = $language;
+
+        if ($this->use_cookie) {
+            set_cookie(self::$context . "_language", $language, time() + 365 * 24 * 3600, ["samesite" => "strict", "path" => $this->cookie_path]);
+        }
     } // setCurrentLanguage
-    
+
+    /**
+     * Returns the fallback language. If set and a translation
+     * is missing on a language, the translation on this language will be used.
+     *
+     * @return string
+     * Returns the fallback language.
+     *
+     * @author Oleg Schildt
+     */
+    public function getFallbackLanguage(): string
+    {
+        return $this->use_fallback_language;
+    } // getFallbackLanguage
+
     /**
      * Returns the current language.
      *
@@ -419,25 +605,27 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    public function getCurrentLanguage()
+    public function getCurrentLanguage(): string
     {
-        return self::$current_language;
+        if (empty(self::$current_language[self::$context])) {
+            foreach (self::$supported_languages as $lng) {
+                self::$current_language[self::$context] = $lng;
+                break;
+            }
+        }
+
+        return self::$current_language[self::$context];
     } // getCurrentLanguage
-    
+
     /**
-     * Provides the text translation for the text ID for the given langauge.
+     * Provides the text translation for the text ID for the given language.
      *
      * @param string $text_id
      * Text ID
      *
      * @param string $lng
-     * The langauge. If it is not specified,
-     * the default langauge is used.
-     *
-     * @param boolean $warn_missing
-     * If it is set to true,
-     * the E_USER_NOTICE is triggered in the case of missing
-     * translations.
+     * The language. If it is not specified,
+     * the default language is used.
      *
      * @param string $default_text
      * The default text to be used if there is no translation.
@@ -446,68 +634,103 @@ class LanguageManager implements ILanguageManager
      * Returns the translation text or the $default_text/$text_id if no translation
      * is found.
      *
+     * @throws \Exception
+     * It might throw exceptions in the case of any errors.
+     *
      * @author Oleg Schildt
      */
-    public function text($text_id, $lng = "", $warn_missing = true, $default_text = "")
+    public function text(string $text_id, string $lng = "", string $default_text = ""): string
     {
         if (empty($lng)) {
             $lng = $this->getCurrentLanguage();
         }
-        
-        if (empty(self::$texts[$lng][$text_id])) {
-            if ($warn_missing) {
-                trigger_error("No translation for the text '$text_id' in the language [$lng]!", E_USER_NOTICE);
+
+        if (!$this->hasTranslation($text_id, $lng)) {
+            if ($this->warn_missing) {
+                trigger_error("No translation for the text '$text_id' in the language [$lng]!", E_USER_WARNING);
             }
-            
+
             if (empty($default_text)) {
+                if (!empty($this->use_fallback_language)) {
+                    return $this->try_text($text_id, $this->use_fallback_language);
+                }
+              
                 return $text_id;
             } else {
                 return $default_text;
             }
         }
-        
+
         return self::$texts[$lng][$text_id];
     } // text
-    
+
     /**
-     * Checks whether the text translation for the text ID for the given langauge exists.
+     * Provides the text translation for the text ID for the given language if the translation exists.
+     * Otherwise, it returns the text ID and emits no warning.
      *
      * @param string $text_id
      * Text ID
      *
      * @param string $lng
-     * The langauge. If it is not specified,
-     * the default langauge is used.
+     * The language. If it is not specified,
+     * the default language is used.
      *
-     * @return boolean
+     * @return string
+     * Returns the translation text or the $default_text/$text_id if no translation
+     * is found.
+     *
+     * @throws \Exception
+     * It might throw exceptions in the case of any errors.
+     *
+     * @author Oleg Schildt
+     */
+    public function try_text(string $text_id, string $lng = ""): string
+    {
+        if (!$this->hasTranslation($text_id, $lng)) {
+            if (!empty($this->use_fallback_language) && $this->hasTranslation($text_id, $this->use_fallback_language)) {
+                return text($text_id, $this->use_fallback_language);
+            }
+            
+            return $text_id;
+        }
+        
+        return text($text_id, $lng);
+    } // try_text
+
+    /**
+     * Checks whether the text translation for the text ID for the given language exists.
+     *
+     * @param string $text_id
+     * Text ID
+     *
+     * @param string $lng
+     * The language. If it is not specified,
+     * the default language is used.
+     *
+     * @return bool
      * Returns true if the translation exists, otherwise false.
      *
      * @author Oleg Schildt
      */
-    public function hasTranslation($text_id, $lng = "")
+    public function hasTranslation(string $text_id, string $lng = ""): bool
     {
         if (empty($lng)) {
             $lng = $this->getCurrentLanguage();
         }
-        
+
         return !empty(self::$texts[$lng][$text_id]);
     } // hasTranslation
-    
+
     /**
      * Provides the text translation for the language name by the code
-     * for the given langauge.
+     * for the given language.
      *
      * @param string $code
      * Language ISO code (lowercase, e.g. en, de, fr).
      *
      * @param string $lng
-     * The langauge. If it is not specified,
-     * the default langauge is used.
-     *
-     * @param boolean $warn_missing
-     * If it is set to true,
-     * the E_USER_NOTICE is triggered in the case of mussing
-     * translations.
+     * The language. If it is not specified,
+     * the default language is used.
      *
      * @return string
      * Returns the translation text for the language name or the $code if no translation
@@ -520,22 +743,22 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    public function getLanguageName($code, $lng = "", $warn_missing = true)
+    public function getLanguageName(string $code, string $lng = ""): string
     {
         if (empty($lng)) {
             $lng = $this->getCurrentLanguage();
         }
-        
+
         if (empty(self::$languages[$lng][$code])) {
-            if ($warn_missing) {
-                trigger_error("No translation for the language name [$code] in the language [$lng]!", E_USER_NOTICE);
+            if ($this->warn_missing) {
+                trigger_error("No translation for the language name [$code] in the language [$lng]!", E_USER_WARNING);
             }
             return $code;
         }
-        
+
         return self::$languages[$lng][$code];
     } // getLanguageName
-    
+
     /**
      * Tries to find the language code by the given name.
      *
@@ -552,23 +775,23 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    public function getLanguageCode($lang_name)
+    public function getLanguageCode(string $lang_name): string
     {
         foreach (self::$supported_languages as $lng) {
             if (empty(self::$languages[$lng])) {
                 continue;
             }
-            
+
             foreach (self::$languages[$lng] as $code => $translation) {
                 if (strcasecmp($lang_name, $translation) == 0) {
                     return $code;
                 }
             } // foreach
         } // foreach
-        
+
         return "";
     } // getLanguageCode
-    
+
     /**
      * Checks whether the language code is valid (has translation).
      *
@@ -576,11 +799,11 @@ class LanguageManager implements ILanguageManager
      * Language ISO code (lowercase, e.g. en, de, fr).
      *
      * @param string $lng
-     * The langauge. If it is not specified,
-     * the default langauge is used.
+     * The language. If it is not specified,
+     * the default language is used.
      *
-     * @return boolean
-     * Returns true if the langauge code is valid (has translation), otherwise false.
+     * @return bool
+     * Returns true if the language code is valid (has translation), otherwise false.
      *
      * @see LanguageManager::getLanguageName()
      * @see LanguageManager::getLanguageCode()
@@ -589,15 +812,15 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    public function validateLanguageCode($code, $lng = "")
+    public function validateLanguageCode(string $code, string $lng = ""): bool
     {
         if (empty($lng)) {
             $lng = $this->getCurrentLanguage();
         }
-        
+
         return !empty(self::$languages[$lng][$code]);
     } // validateLanguageCode
-    
+
     /**
      * Provides the list of languages for the given language in the form "code" => "translation".
      *
@@ -605,14 +828,14 @@ class LanguageManager implements ILanguageManager
      * Target array where the language list should be loaded.
      *
      * @param string $lng
-     * The langauge. If it is not specified,
-     * the default langauge is used.
+     * The language. If it is not specified,
+     * the default language is used.
      *
      * @param array $display_first
      * List of the language codes to be displayed first in the order, they appear in the list.
      *
-     * @return boolean
-     * Returns true if the langauge list is successfully retrieved, otherwise false.
+     * @return bool
+     * Returns true if the language list is successfully retrieved, otherwise false.
      *
      * @see LanguageManager::getLanguageName()
      * @see LanguageManager::getLanguageCode()
@@ -621,42 +844,37 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    public function getLanguageList(&$language_list, $lng = "", $display_first = [])
+    public function getLanguageList(array &$language_list, string $lng = "", array $display_first = []): bool
     {
         if (empty($lng)) {
             $lng = $this->getCurrentLanguage();
         }
-        
+
         if (empty(self::$languages[$lng])) {
             return false;
         }
-    
+
         $language_list = array_flip($display_first);
-    
+
         asort(self::$languages[$lng], SORT_LOCALE_STRING);
-    
-        foreach(self::$languages[$lng] as $code => $name) {
+
+        foreach (self::$languages[$lng] as $code => $name) {
             $language_list[$code] = $name;
         }
 
         return true;
     } // getLanguageList
-    
+
     /**
      * Provides the text translation for the country name by the code
-     * for the given langauge.
+     * for the given language.
      *
      * @param string $code
      * Country ISO code (uppercase, e.g. US, DE, FR).
      *
      * @param string $lng
-     * The langauge. If it is not specified,
-     * the default langauge is used.
-     *
-     * @param boolean $warn_missing
-     * If it is set to true,
-     * the E_USER_NOTICE is triggered in the case of mussing
-     * translations.
+     * The language. If it is not specified,
+     * the default language is used.
      *
      * @return string
      * Returns the translation text for the country name or the $code if no translation
@@ -669,22 +887,22 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    public function getCountryName($code, $lng = "", $warn_missing = true)
+    public function getCountryName(string $code, string $lng = ""): string
     {
         if (empty($lng)) {
             $lng = $this->getCurrentLanguage();
         }
-        
+
         if (empty(self::$countries[$lng][$code])) {
-            if ($warn_missing) {
-                trigger_error("No translation for the country name [$code] in the language [$lng]!", E_USER_NOTICE);
+            if ($this->warn_missing) {
+                trigger_error("No translation for the country name [$code] in the language [$lng]!", E_USER_WARNING);
             }
             return $code;
         }
-        
+
         return self::$countries[$lng][$code];
     } // getCountryName
-    
+
     /**
      * Tries to find the country code by the given name.
      *
@@ -701,23 +919,23 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    public function getCountryCode($country_name)
+    public function getCountryCode(string $country_name): string
     {
         foreach (self::$supported_languages as $lng) {
             if (empty(self::$countries[$lng])) {
                 continue;
             }
-            
+
             foreach (self::$countries[$lng] as $code => $translation) {
                 if (strcasecmp($country_name, $translation) == 0) {
                     return $code;
                 }
             } // foreach
         } // foreach
-        
+
         return "";
     } // getCountryCode
-    
+
     /**
      * Checks whether the country code is valid (has translation).
      *
@@ -725,10 +943,10 @@ class LanguageManager implements ILanguageManager
      * Country ISO code (uppercase, e.g. US, DE, FR).
      *
      * @param string $lng
-     * The langauge. If it is not specified,
-     * the default langauge is used.
+     * The language. If it is not specified,
+     * the default language is used.
      *
-     * @return boolean
+     * @return bool
      * Returns true if the country code is valid (has translation), otherwise false.
      *
      * @see LanguageManager::getCountryName()
@@ -738,15 +956,15 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    public function validateCountryCode($code, $lng = "")
+    public function validateCountryCode(string $code, string $lng = ""): bool
     {
         if (empty($lng)) {
             $lng = $this->getCurrentLanguage();
         }
-        
+
         return !empty(self::$countries[$lng][$code]);
     } // validateCountryCode
-    
+
     /**
      * Provides the list of countries for the given language in the form "code" => "translation".
      *
@@ -754,13 +972,13 @@ class LanguageManager implements ILanguageManager
      * Target array where the country list should be loaded.
      *
      * @param string $lng
-     * The langauge. If it is not specified,
-     * the default langauge is used.
+     * The language. If it is not specified,
+     * the default language is used.
      *
      * @param array $display_first
      * List of the country codes to be displayed first in the order, they appear in the list.
      *
-     * @return boolean
+     * @return bool
      * Returns true if the country list is successfully retrieved, otherwise false.
      *
      * @see LanguageManager::getCountryName()
@@ -770,24 +988,24 @@ class LanguageManager implements ILanguageManager
      *
      * @author Oleg Schildt
      */
-    public function getCountryList(&$country_list, $lng = "", $display_first = [])
+    public function getCountryList(array &$country_list, string $lng = "", array $display_first = []): bool
     {
         if (empty($lng)) {
             $lng = $this->getCurrentLanguage();
         }
-        
+
         if (empty(self::$countries[$lng])) {
             return false;
         }
-    
+
         $country_list = array_flip($display_first);
 
         asort(self::$countries[$lng], SORT_LOCALE_STRING);
-        
-        foreach(self::$countries[$lng] as $code => $name) {
+
+        foreach (self::$countries[$lng] as $code => $name) {
             $country_list[$code] = $name;
         }
-        
+
         return true;
     } // getCountryList
 } // LanguageManager
