@@ -27,39 +27,39 @@ class ObjectFactory
      *
      * @author Oleg Schildt
      */
-    static protected $singletons = [];
+    static protected array $singletons = [];
     
     /**
-     * Internal array for storing the mapping betwen interfaces/classes and the bound classes.
+     * Internal array for storing the mapping between interfaces/classes and the bound classes.
      *
      * @var array
      *
      * @author Oleg Schildt
      */
-    static protected $itable = [];
+    static protected array $itable = [];
     
     /**
      * Binds a class to an interface or a parent class.
      *
      * The key point of this approach is the definition common interfaces
-     * and implementation of them in classes. When an object that supoorts an
+     * and implementation of them in classes. When an object that supports an
      * interface is requested, an instance of the corresponding bound class
      * is created. When you want to change the class, you need just bind
      * the new class to the interface.
      *
-     * You can also bind a class to itself and reqest it by own name in factory
+     * You can also bind a class to itself and request it by own name in factory
      * method. And later, you can implement a derived class and change the binding
      * without affecting the code in the business logic.
      *
-     * @param string|object $interface_or_class
+     * @param object|string $interface_or_class
      * Name of the class/interface as string or the class/interface.
      *
-     * @param string|object $class
+     * @param object|string $class
      * The class which instance should be created if the object is requested.
      *
-     * @param callable $init_function
+     * @param ?callable $init_function
      * The optional initialization function. You can provide it to do some
-     * custom intialization. The signature of
+     * custom initialization. The signature of
      * this function is:
      *
      * ```php
@@ -94,63 +94,129 @@ class ObjectFactory
      *
      * @author Oleg Schildt
      */
-    static public function bindClass($interface_or_class, $class, $init_function = null)
+    static public function bindClass(object|string $interface_or_class, object|string $class, ?callable $init_function = null): void
+    {
+        self::bindClassContext("default", $interface_or_class, $class, $init_function);
+    } // bindClass
+
+    /**
+     * Binds a class to an interface or a parent class.
+     *
+     * The key point of this approach is the definition common interfaces
+     * and implementation of them in classes. When an object that supports an
+     * interface is requested, an instance of the corresponding bound class
+     * is created. When you want to change the class, you need just bind
+     * the new class to the interface.
+     *
+     * You can also bind a class to itself and request it by own name in factory
+     * method. And later, you can implement a derived class and change the binding
+     * without affecting the code in the business logic.
+     *
+     * @param string $context
+     * The context of binding. You can introduce a different context, and bind another implementation.
+     * Then, you can request another instance specifying the context explicitly.
+     *
+     * @param object|string $interface_or_class
+     * Name of the class/interface as string or the class/interface.
+     *
+     * @param object|string $class
+     * The class which instance should be created if the object is requested.
+     *
+     * @param ?callable $init_function
+     * The optional initialization function. You can provide it to do some
+     * custom initialization. The signature of
+     * this function is:
+     *
+     * ```php
+     * function (object $instance) : void;
+     * ```
+     *
+     * - $instance - the created instance.
+     *
+     * Example:
+     *
+     * ```php
+     * ObjectFactory::bindClassContext("console", ILanguageManager::class, LanguageManager::class, function($instance) {
+     *   $instance->detectLanguage();
+     * });
+     *
+     * ObjectFactory::bindClassContext("console", IRecordsetManager::class, RecordsetManager::class, function($instance) {
+     *   $instance->setDBWorker(dbworker());
+     * });
+     * ```
+     *
+     * @return void
+     *
+     * @throws \Exception
+     * It might throw the following exceptions in the case of any errors:
+     *
+     * - if the interface or bound class is not specified.
+     * - if the interface or class does not exist.
+     * - if the bound class is empty.
+     * - if the bound class does not implement the corresponding interface.
+     * - if the bound class is not instantiable.
+     * - if the check of the classes and interfaces fails.
+     *
+     * @author Oleg Schildt
+     */
+    static public function bindClassContext(string $context, object|string $interface_or_class, object|string $class, ?callable $init_function = null): void
     {
         if (empty($class)) {
             throw new \Exception("Bound class is empty!");
         }
-        
+
         if (empty($interface_or_class)) {
             throw new \Exception("Bound interface or class is empty!");
         }
-        
+
         if (!interface_exists($interface_or_class) && !class_exists($interface_or_class)) {
             throw new \Exception(sprintf("The interface or class '%s' does not exist!", $interface_or_class));
         }
-        
+
         if (!class_exists($class)) {
             throw new \Exception(sprintf("The class '%s' does not exist!", $class));
         }
-        
+
         $ic = new \ReflectionClass($interface_or_class);
         $c = new \ReflectionClass($class);
-        
+
         if (!$c->isInstantiable()) {
             throw new \Exception(sprintf("The class '%s' is not instantiable!", $c->getName()));
         }
-        
+
         if ($c != $ic) {
             if (!$c->isSubclassOf($ic)) {
                 throw new \Exception(sprintf("The class '%s' does not implement the interface '%s'!", $c->getName(), $ic->getName()));
             }
         }
-        
+
         $f = null;
         if ($init_function !== null) {
             if (!is_callable($init_function)) {
                 throw new \Exception(sprintf("'%s' is not a function!", $init_function));
             }
-            
+
             $f = new \ReflectionFunction($init_function);
         }
-        
-        self::$itable[$ic->getName()] = ["class" => $c, "init_function" => $f];
-    } // bindClass
-    
+
+        self::$itable[$context][$ic->getName()] = ["class" => $c, "init_function" => $f];
+    } // bindClassContext
+
     /**
      * Сreates an object that support the interface $interface_or_class.
      *
-     * @param string|object $interface_or_class
+     * @param object|string $interface_or_class
      * Name of the class/interface which instance should be created.
      *
-     * @param boolean $singleton
+     * @param bool $singleton
      * If the parameter is true, it ensures that only one instance of this object exists.
      * The singleton is a usual patter for the action objects like SessionManager, EventManager,
      * DBWorker etc. It makes no sense to produce many instances of such classes,
      * it wastes the computer resources and might cause errors.
      *
-     * If the parameter is false, then, by each request, a new object is created. If you request
-     * data objects like User, a separate instance must be created for each item.
+     * @param string $context
+     * The context of binding. You can introduce a different context, and bind another implementation.
+     * Then, you can request another instance specifying the context explicitly.
      *
      * @return object
      * Returns object of the class bound to the interface.
@@ -167,7 +233,7 @@ class ObjectFactory
      *
      * @author Oleg Schildt
      */
-    static public function getInstance($interface_or_class, $singleton)
+    static public function getInstance(object|string $interface_or_class, bool $singleton, string $context = "default"): object
     {
         if (empty($interface_or_class)) {
             throw new \Exception("Class or interface is not specified!");
@@ -181,30 +247,30 @@ class ObjectFactory
         
         $class_name = $class->getName();
         
-        if (empty(self::$itable[$class_name])) {
-            throw new \Exception(sprintf("The interface or class '%s' has no bound class!", $class_name));
+        if (empty(self::$itable[$context][$class_name])) {
+            throw new \Exception(sprintf("The interface or class '%s' has no bound class in the context '$context'!", $class_name));
         }
         
         // if not singleton, we create a new instance every time it is requested
         if (!$singleton) {
-            $instance = self::$itable[$class_name]["class"]->newInstance();
+            $instance = self::$itable[$context][$class_name]["class"]->newInstance();
             
-            if (!empty(self::$itable[$class_name]["init_function"])) {
-                self::$itable[$class_name]["init_function"]->invoke($instance);
+            if (!empty(self::$itable[$context][$class_name]["init_function"])) {
+                self::$itable[$context][$class_name]["init_function"]->invoke($instance);
             }
             
             return $instance;
         }
         
         // if singleton, we create an instance only if it does not exist yet
-        if (empty(self::$singletons[$class_name])) {
-            self::$singletons[$class_name] = self::$itable[$class_name]["class"]->newInstance();
+        if (empty(self::$singletons[$context][$class_name])) {
+            self::$singletons[$context][$class_name] = self::$itable[$context][$class_name]["class"]->newInstance();
             
-            if (!empty(self::$itable[$class_name]["init_function"])) {
-                self::$itable[$class_name]["init_function"]->invoke(self::$singletons[$class_name]);
+            if (!empty(self::$itable[$context][$class_name]["init_function"])) {
+                self::$itable[$context][$class_name]["init_function"]->invoke(self::$singletons[$context][$class_name]);
             }
         }
         
-        return self::$singletons[$class_name];
+        return self::$singletons[$context][$class_name];
     } // getInstance
 } // ObjectFactory
