@@ -11,8 +11,9 @@
 
 namespace SmartFactory;
 
-use SmartFactory\Interfaces\ISettingsManager;
-use SmartFactory\DatabaseWorkers\DBWorker;
+use \SmartFactory\Interfaces\ISettingsManager;
+use \SmartFactory\DatabaseWorkers\DBWorker;
+use SmartFactory\Interfaces\ISettingsValidator;
 
 /**
  * Class for management of the application settings.
@@ -38,11 +39,11 @@ class RuntimeSettingsManager implements ISettingsManager
     /**
      * Internal variable for storing the dbworker.
      *
-     * @var DatabaseWorkers\DBWorker
+     * @var ?DatabaseWorkers\DBWorker
      *
      * @author Oleg Schildt
      */
-    protected $dbworker;
+    protected ?DatabaseWorkers\DBWorker $dbworker = null;
     
     /**
      * Internal variable for storing the target settings table name.
@@ -51,7 +52,7 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    protected $settings_table;
+    protected string $settings_table;
     
     /**
      * Internal variable for storing the target column.
@@ -60,7 +61,7 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    protected $settings_column;
+    protected string $settings_column;
     
     /**
      * Internal variable for storing the current context.
@@ -72,19 +73,19 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    protected $context = "default";
+    protected string $context = "default";
     
     /**
      * Internal variable for storing the validator.
      *
-     * @var Interfaces\ISettingsValidator
+     * @var ?Interfaces\ISettingsValidator
      *
      * @see RuntimeSettingsManager::getValidator()
      * @see RuntimeSettingsManager::setValidator()
      *
      * @author Oleg Schildt
      */
-    protected $validator = null;
+    protected ?Interfaces\ISettingsValidator $validator = null;
     
     /**
      * Internal variable for storing the array of the settings values.
@@ -93,14 +94,13 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    protected $settings = [];
+    protected array $settings = [];
     
     /**
      * This is internal auxiliary function for checking that the settings
-     * manager is intialized correctly.
+     * manager is initialized correctly.
      *
-     * @return boolean
-     * It should return true if the settings manager is intialized correctly, otherwise false.
+     * @return void
      *
      * @throws \Exception
      * It might throw an exception in the case of any errors:
@@ -110,7 +110,7 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    protected function validateParameters()
+    protected function validateParameters(): void
     {
         if (empty($this->dbworker)) {
             throw new \Exception("The 'dbworker' is not specified!");
@@ -127,19 +127,16 @@ class RuntimeSettingsManager implements ISettingsManager
         if (empty($this->settings_column)) {
             throw new \Exception("The 'settings_column' is not specified!");
         }
-        
-        return true;
     } // validateParameters
     
     /**
      * This is internal auxiliary function for converting the settings to JSON and storing it
-     * to the target table defined by the iniailization.
+     * to the target table defined by the initialization.
      *
      * @param array &$data
      * The array with the settings values to be saved.
      *
-     * @return boolean
-     * Returns true if the data has been successfully saved, otherwise false.
+     * @return void
      *
      * @throws \Exception
      * It might throw an exception in the case of any errors:
@@ -152,7 +149,7 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    protected function saveJSON(&$data)
+    protected function saveJSON(array &$data): void
     {
         $this->validateParameters();
         
@@ -175,26 +172,23 @@ class RuntimeSettingsManager implements ISettingsManager
         $json = $this->dbworker->escape($json);
         
         if ($must_insert) {
-            $query = "INSERT INTO " . $this->settings_table . "(" . $this->settings_column . ")\n";
-            $query .= "VALUES ('$json')";
+            $query = "insert into " . $this->settings_table . "(" . $this->settings_column . ")\n";
+            $query .= "values ('$json')";
         } else {
-            $query = "UPDATE " . $this->settings_table . " SET " . $this->settings_column . " = '$json'";
+            $query = "update " . $this->settings_table . " set " . $this->settings_column . " = '$json'";
         }
         
         $this->dbworker->execute_query($query);
-        
-        return true;
     } // saveJSON
     
     /**
      * This is internal auxiliary function for loading the settings from the target table
-     * defined by the iniailization.
+     * defined by the initialization.
      *
      * @param array &$data
      * The target array with the settings values to be loaded.
      *
-     * @return boolean
-     * Returns true if the data has been successfully loaded, otherwise false.
+     * @return void
      *
      * @throws \Exception
      * It might throw an exception in the case of any errors:
@@ -208,11 +202,11 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    protected function loadJSON(&$data)
+    protected function loadJSON(array &$data): void
     {
         $this->validateParameters();
         
-        $query = "SELECT " . $this->settings_column . " FROM " . $this->settings_table;
+        $query = "select " . $this->settings_column . " from " . $this->settings_table;
         
         $this->dbworker->execute_query($query);
         
@@ -225,7 +219,7 @@ class RuntimeSettingsManager implements ISettingsManager
         $this->dbworker->free_result();
         
         if (empty($json)) {
-            return true;
+            return;
         }
         
         try {
@@ -233,8 +227,6 @@ class RuntimeSettingsManager implements ISettingsManager
         } catch (\Throwable $ex) {
             throw new \Exception("JSON parse error: " . $ex->getMessage());
         }
-        
-        return true;
     } // loadJSON
     
     /**
@@ -249,8 +241,7 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * - $parameters["settings_column"] - the name of the column for the storing of the settings.
      *
-     * @return boolean
-     * Returns true upon successful initialization, otherwise false.
+     * @return void
      *
      * @throws \Exception
      * It might throw an exception in the case of any errors:
@@ -260,7 +251,7 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    public function init($parameters)
+    public function init(array $parameters): void
     {
         if (!empty($parameters["dbworker"])) {
             $this->dbworker = $parameters["dbworker"];
@@ -274,13 +265,13 @@ class RuntimeSettingsManager implements ISettingsManager
             $this->settings_column = $parameters["settings_column"];
         }
         
-        return $this->validateParameters();
+        $this->validateParameters();
     } // init
     
     /**
      * Sets the validator for the settings.
      *
-     * @param Interfaces\ISettingsValidator $validator
+     * @param ISettingsValidator $validator
      * The settings validator.
      *
      * @return void
@@ -290,7 +281,7 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    public function setValidator($validator)
+    public function setValidator(ISettingsValidator $validator): void
     {
         $this->validator = $validator;
     } // setValidator
@@ -298,7 +289,7 @@ class RuntimeSettingsManager implements ISettingsManager
     /**
      * Returns the validator for the settings.
      *
-     * @return Interfaces\ISettingsValidator|null
+     * @return ?Interfaces\ISettingsValidator
      * Returns the validator for the settings or null if none is defined.
      *
      * @see RuntimeSettingsManager::setValidator()
@@ -306,7 +297,7 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    public function getValidator()
+    public function getValidator(): ?Interfaces\ISettingsValidator
     {
         return $this->validator;
     } // getValidator
@@ -332,7 +323,7 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    public function setContext($context = "default")
+    public function setContext(string $context = "default"): void
     {
         $this->context = $context;
     } // setContext
@@ -356,7 +347,7 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    public function getContext()
+    public function getContext(): string
     {
         return $this->context;
     } // getContext
@@ -384,7 +375,7 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    public function setParameter($name, $value)
+    public function setParameter(string $name, mixed $value): void
     {
         if (empty($this->settings)) {
             $this->loadSettings();
@@ -397,10 +388,10 @@ class RuntimeSettingsManager implements ISettingsManager
     /**
      * Sets settings parameters from an array.
      *
-     * @param array &$parameters
+     * @param array $parameters
      * Array of parameters in the form key => value.
      *
-     * @param boolean $force_creation
+     * @param bool $force_creation
      * Flag which defines whether the parameter should be created
      * if not exists. If false, only existing parameters are updated.
      *
@@ -419,7 +410,7 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    public function setParameters(&$parameters, $force_creation = false)
+    public function setParameters(array $parameters, bool $force_creation = false): void
     {
         if (empty($this->settings)) {
             $this->loadSettings();
@@ -440,9 +431,9 @@ class RuntimeSettingsManager implements ISettingsManager
      * @param string $name
      * The name of the settings parameter.
      *
-     * @param mixed $default
+     * @param mixed|null $default
      * The default value of the settings parameter if it is not set yet.
-     * The parameter is a confortable way to pre-set a parameter
+     * The parameter is a conformable way to pre-set a parameter
      * to a default value if its value is not set yet.
      *
      * @return mixed
@@ -460,7 +451,7 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    public function getParameter($name, $default = null)
+    public function getParameter(string $name, mixed $default = null): mixed
     {
         if (empty($this->settings)) {
             $this->loadSettings();
@@ -479,9 +470,10 @@ class RuntimeSettingsManager implements ISettingsManager
      * It should be called after settings the new values of the parameters
      * and before their saving.
      *
-     * @return boolean
-     * Returns true if there is no validator defined, otherwise lets
-     * the validator validate the settings.
+     * @return void
+     *
+     * @throws \Exception
+     * It might throw exceptions in the case of any errors.
      *
      * @uses \SmartFactory\Interfaces\ISettingsValidator
      *
@@ -490,20 +482,19 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    public function validateSettings()
+    public function validateSettings(): void
     {
         if (empty($this->validator)) {
-            return true;
+            return;
         }
         
-        return $this->validator->validate($this, $this->context);
+        $this->validator->validate($this, $this->context);
     } // validateSettings
     
     /**
      * Loads the settings from the target table.
      *
-     * @return boolean
-     * Returns true if the settings have been successfully loaded, otherwise false.
+     * @return void
      *
      * @throws \Exception
      * It might throw an exception in the case of any errors:
@@ -516,16 +507,15 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    public function loadSettings()
+    public function loadSettings(): void
     {
-        return $this->loadJSON($this->settings);
+        $this->loadJSON($this->settings);
     } // loadSettings
     
     /**
      * Saves the settings from to the target table.
      *
-     * @return boolean
-     * Returns true if the settings have been successfully saved, otherwise false.
+     * @return void
      *
      * @throws \Exception
      * It might throw an exception in the case of any errors:
@@ -538,12 +528,12 @@ class RuntimeSettingsManager implements ISettingsManager
      *
      * @author Oleg Schildt
      */
-    public function saveSettings()
+    public function saveSettings(): void
     {
         if (empty($this->settings)) {
             $this->loadSettings();
         }
 
-        return $this->saveJSON($this->settings);
+        $this->saveJSON($this->settings);
     } // saveSettings
 } // RuntimeSettingsManager
