@@ -1,5 +1,6 @@
 var ApiGen = ApiGen || {};
 ApiGen.elements = [
+{"file":"global-$bogo.html","label":"$bogo - Global variable"},
 {"file":"package-Database.html","label":"Database - Package"},
 {"file":"package-Factory.html","label":"Factory - Package"},
 {"file":"package-HTML Utils.html","label":"HTML Utils - Package"},
@@ -33,26 +34,37 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html","label":"DBWorker - Class"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#$db_name","label":"DBWorker::$db_name"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#$db_password","label":"DBWorker::$db_password"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#$db_port","label":"DBWorker::$db_port"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#$db_server","label":"DBWorker::$db_server"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#$db_user","label":"DBWorker::$db_user"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#$is_clone","label":"DBWorker::$is_clone"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#$last_query","label":"DBWorker::$last_query"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#$logging","label":"DBWorker::$logging"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#DB_AS_IS","label":"DBWorker::DB_AS_IS - Constant"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#DB_DATE","label":"DBWorker::DB_DATE - Constant"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#DB_DATETIME","label":"DBWorker::DB_DATETIME - Constant"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#DB_GEOMETRY","label":"DBWorker::DB_GEOMETRY - Constant"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#DB_GEOMETRY_4326","label":"DBWorker::DB_GEOMETRY_4326 - Constant"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#DB_LARGE_OBJECT_STREAM","label":"DBWorker::DB_LARGE_OBJECT_STREAM - Constant"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#DB_NUMBER","label":"DBWorker::DB_NUMBER - Constant"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#DB_STRING","label":"DBWorker::DB_STRING - Constant"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#ERR_CONNECTION_DATA_INCOMPLETE","label":"DBWorker::ERR_CONNECTION_DATA_INCOMPLETE - Constant"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#ERR_CONNECTION_FAILED","label":"DBWorker::ERR_CONNECTION_FAILED - Constant"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#ERR_DATABASE_NOT_FOUND","label":"DBWorker::ERR_DATABASE_NOT_FOUND - Constant"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#ERR_HOST_UNREACHABLE","label":"DBWorker::ERR_HOST_UNREACHABLE - Constant"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#ERR_NOT_CONNECTED","label":"DBWorker::ERR_NOT_CONNECTED - Constant"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#ERR_QUERY_FAILED","label":"DBWorker::ERR_QUERY_FAILED - Constant"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#ERR_STREAM_ERROR","label":"DBWorker::ERR_STREAM_ERROR - Constant"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#ERR_WRONG_DATA_FORMAT","label":"DBWorker::ERR_WRONG_DATA_FORMAT - Constant"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#ERR_WRONG_USER_CREDENTIALS","label":"DBWorker::ERR_WRONG_USER_CREDENTIALS - Constant"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#affected_count","label":"DBWorker::affected_count()"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#build_select_query","label":"DBWorker::build_select_query()"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#check_connection","label":"DBWorker::check_connection()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#close_connection","label":"DBWorker::close_connection()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#commit_transaction","label":"DBWorker::commit_transaction()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#connect","label":"DBWorker::connect()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#create_clone","label":"DBWorker::create_clone()"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#enable_logging","label":"DBWorker::enable_logging()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#escape","label":"DBWorker::escape()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#execute_prepared_query","label":"DBWorker::execute_prepared_query()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#execute_procedure","label":"DBWorker::execute_procedure()"},
@@ -77,13 +89,17 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#insert_id","label":"DBWorker::insert_id()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#is_connected","label":"DBWorker::is_connected()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#is_extension_installed","label":"DBWorker::is_extension_installed()"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#number_or_null","label":"DBWorker::number_or_null()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#prepare_for_query","label":"DBWorker::prepare_for_query()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#prepare_query","label":"DBWorker::prepare_query()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#qualify_name_with_schema","label":"DBWorker::qualify_name_with_schema()"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#quotes_or_null","label":"DBWorker::quotes_or_null()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#rollback_transaction","label":"DBWorker::rollback_transaction()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#start_transaction","label":"DBWorker::start_transaction()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#stream_long_data","label":"DBWorker::stream_long_data()"},
 {"file":"class-SmartFactory.DatabaseWorkers.DBWorker.html#use_database","label":"DBWorker::use_database()"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorkerException.html","label":"DBWorkerException - Class"},
+{"file":"class-SmartFactory.DatabaseWorkers.DBWorkerException.html#__construct","label":"DBWorkerException::__construct()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MSSQL_DBWorker.html","label":"MSSQL_DBWorker - Class"},
 {"file":"class-SmartFactory.DatabaseWorkers.MSSQL_DBWorker.html#$connection","label":"MSSQL_DBWorker::$connection"},
 {"file":"class-SmartFactory.DatabaseWorkers.MSSQL_DBWorker.html#$field_names","label":"MSSQL_DBWorker::$field_names"},
@@ -95,6 +111,7 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.DatabaseWorkers.MSSQL_DBWorker.html#__construct","label":"MSSQL_DBWorker::__construct()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MSSQL_DBWorker.html#__destruct","label":"MSSQL_DBWorker::__destruct()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MSSQL_DBWorker.html#affected_count","label":"MSSQL_DBWorker::affected_count()"},
+{"file":"class-SmartFactory.DatabaseWorkers.MSSQL_DBWorker.html#build_select_query","label":"MSSQL_DBWorker::build_select_query()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MSSQL_DBWorker.html#close_connection","label":"MSSQL_DBWorker::close_connection()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MSSQL_DBWorker.html#commit_transaction","label":"MSSQL_DBWorker::commit_transaction()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MSSQL_DBWorker.html#connect","label":"MSSQL_DBWorker::connect()"},
@@ -131,6 +148,7 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.DatabaseWorkers.MSSQL_DBWorker.html#sys_get_errors","label":"MSSQL_DBWorker::sys_get_errors()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MSSQL_DBWorker.html#use_database","label":"MSSQL_DBWorker::use_database()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html","label":"MySQL_DBWorker - Class"},
+{"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#$connection","label":"MySQL_DBWorker::$connection"},
 {"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#$field_names","label":"MySQL_DBWorker::$field_names"},
 {"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#$mysqli","label":"MySQL_DBWorker::$mysqli"},
 {"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#$mysqli_result","label":"MySQL_DBWorker::$mysqli_result"},
@@ -141,6 +159,7 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#__construct","label":"MySQL_DBWorker::__construct()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#__destruct","label":"MySQL_DBWorker::__destruct()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#affected_count","label":"MySQL_DBWorker::affected_count()"},
+{"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#build_select_query","label":"MySQL_DBWorker::build_select_query()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#close_connection","label":"MySQL_DBWorker::close_connection()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#commit_transaction","label":"MySQL_DBWorker::commit_transaction()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#connect","label":"MySQL_DBWorker::connect()"},
@@ -175,6 +194,54 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#start_transaction","label":"MySQL_DBWorker::start_transaction()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#stream_long_data","label":"MySQL_DBWorker::stream_long_data()"},
 {"file":"class-SmartFactory.DatabaseWorkers.MySQL_DBWorker.html#use_database","label":"MySQL_DBWorker::use_database()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html","label":"PostgreSQL_DBWorker - Class"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#$connection","label":"PostgreSQL_DBWorker::$connection"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#$field_names","label":"PostgreSQL_DBWorker::$field_names"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#$prepared_query","label":"PostgreSQL_DBWorker::$prepared_query"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#$read_only","label":"PostgreSQL_DBWorker::$read_only"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#$result","label":"PostgreSQL_DBWorker::$result"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#$row","label":"PostgreSQL_DBWorker::$row"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#$statement","label":"PostgreSQL_DBWorker::$statement"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#__construct","label":"PostgreSQL_DBWorker::__construct()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#__destruct","label":"PostgreSQL_DBWorker::__destruct()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#affected_count","label":"PostgreSQL_DBWorker::affected_count()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#build_select_query","label":"PostgreSQL_DBWorker::build_select_query()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#close_connection","label":"PostgreSQL_DBWorker::close_connection()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#commit_transaction","label":"PostgreSQL_DBWorker::commit_transaction()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#connect","label":"PostgreSQL_DBWorker::connect()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#create_clone","label":"PostgreSQL_DBWorker::create_clone()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#escape","label":"PostgreSQL_DBWorker::escape()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#execute_prepared_query","label":"PostgreSQL_DBWorker::execute_prepared_query()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#execute_procedure","label":"PostgreSQL_DBWorker::execute_procedure()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#execute_query","label":"PostgreSQL_DBWorker::execute_query()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#fetch_array","label":"PostgreSQL_DBWorker::fetch_array()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#fetch_row","label":"PostgreSQL_DBWorker::fetch_row()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#fetched_count","label":"PostgreSQL_DBWorker::fetched_count()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#field_by_name","label":"PostgreSQL_DBWorker::field_by_name()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#field_by_num","label":"PostgreSQL_DBWorker::field_by_num()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#field_count","label":"PostgreSQL_DBWorker::field_count()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#field_info_by_num","label":"PostgreSQL_DBWorker::field_info_by_num()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#field_name","label":"PostgreSQL_DBWorker::field_name()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#format_date","label":"PostgreSQL_DBWorker::format_date()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#format_datetime","label":"PostgreSQL_DBWorker::format_datetime()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#free_prepared_query","label":"PostgreSQL_DBWorker::free_prepared_query()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#free_result","label":"PostgreSQL_DBWorker::free_result()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#get_extension_name","label":"PostgreSQL_DBWorker::get_extension_name()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#get_rdbms_name","label":"PostgreSQL_DBWorker::get_rdbms_name()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#get_schema","label":"PostgreSQL_DBWorker::get_schema()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#init","label":"PostgreSQL_DBWorker::init()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#insert_id","label":"PostgreSQL_DBWorker::insert_id()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#is_connected","label":"PostgreSQL_DBWorker::is_connected()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#is_extension_installed","label":"PostgreSQL_DBWorker::is_extension_installed()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#prepare_for_query","label":"PostgreSQL_DBWorker::prepare_for_query()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#prepare_query","label":"PostgreSQL_DBWorker::prepare_query()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#qualify_name_with_schema","label":"PostgreSQL_DBWorker::qualify_name_with_schema()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#read_large_object","label":"PostgreSQL_DBWorker::read_large_object()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#rollback_transaction","label":"PostgreSQL_DBWorker::rollback_transaction()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#start_transaction","label":"PostgreSQL_DBWorker::start_transaction()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#stream_long_data","label":"PostgreSQL_DBWorker::stream_long_data()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#sys_get_errors","label":"PostgreSQL_DBWorker::sys_get_errors()"},
+{"file":"class-SmartFactory.DatabaseWorkers.PostgreSQL_DBWorker.html#use_database","label":"PostgreSQL_DBWorker::use_database()"},
 {"file":"class-SmartFactory.DatabaseWorkers.ShardManager.html","label":"ShardManager - Class"},
 {"file":"class-SmartFactory.DatabaseWorkers.ShardManager.html#$load_balancing_groups","label":"ShardManager::$load_balancing_groups"},
 {"file":"class-SmartFactory.DatabaseWorkers.ShardManager.html#$shard_table","label":"ShardManager::$shard_table"},
@@ -184,13 +251,16 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.DebugProfiler.html","label":"DebugProfiler - Class"},
 {"file":"class-SmartFactory.DebugProfiler.html#$log_path","label":"DebugProfiler::$log_path"},
 {"file":"class-SmartFactory.DebugProfiler.html#$profile_time","label":"DebugProfiler::$profile_time"},
+{"file":"class-SmartFactory.DebugProfiler.html#$write_source_file_and_line_by_debug","label":"DebugProfiler::$write_source_file_and_line_by_debug"},
 {"file":"class-SmartFactory.DebugProfiler.html#clearLogFile","label":"DebugProfiler::clearLogFile()"},
 {"file":"class-SmartFactory.DebugProfiler.html#clearLogFiles","label":"DebugProfiler::clearLogFiles()"},
 {"file":"class-SmartFactory.DebugProfiler.html#debugMessage","label":"DebugProfiler::debugMessage()"},
+{"file":"class-SmartFactory.DebugProfiler.html#enableFileAndLineDetails","label":"DebugProfiler::enableFileAndLineDetails()"},
+{"file":"class-SmartFactory.DebugProfiler.html#extract_call_stack","label":"DebugProfiler::extract_call_stack()"},
 {"file":"class-SmartFactory.DebugProfiler.html#fixProfilePoint","label":"DebugProfiler::fixProfilePoint()"},
 {"file":"class-SmartFactory.DebugProfiler.html#init","label":"DebugProfiler::init()"},
-{"file":"class-SmartFactory.DebugProfiler.html#logMessageToFile","label":"DebugProfiler::logMessageToFile()"},
 {"file":"class-SmartFactory.DebugProfiler.html#startProfilePoint","label":"DebugProfiler::startProfilePoint()"},
+{"file":"class-SmartFactory.DebugProfiler.html#trim_path","label":"DebugProfiler::trim_path()"},
 {"file":"class-SmartFactory.ErrorHandler.html","label":"ErrorHandler - Class"},
 {"file":"class-SmartFactory.ErrorHandler.html#$last_error","label":"ErrorHandler::$last_error"},
 {"file":"class-SmartFactory.ErrorHandler.html#$log_path","label":"ErrorHandler::$log_path"},
@@ -209,6 +279,7 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.ErrorHandler.html#traceActive","label":"ErrorHandler::traceActive()"},
 {"file":"class-SmartFactory.ErrorHandler.html#trace_message","label":"ErrorHandler::trace_message()"},
 {"file":"class-SmartFactory.ErrorHandler.html#trim_path","label":"ErrorHandler::trim_path()"},
+{"file":"class-SmartFactory.ErrorHandler.html#truncate_argument","label":"ErrorHandler::truncate_argument()"},
 {"file":"class-SmartFactory.EventManager.html","label":"EventManager - Class"},
 {"file":"class-SmartFactory.EventManager.html#$event_table","label":"EventManager::$event_table"},
 {"file":"class-SmartFactory.EventManager.html#$suspended_events","label":"EventManager::$suspended_events"},
@@ -227,7 +298,6 @@ ApiGen.elements = [
 {"file":"interface-SmartFactory.Interfaces.IDebugProfiler.html#debugMessage","label":"IDebugProfiler::debugMessage()"},
 {"file":"interface-SmartFactory.Interfaces.IDebugProfiler.html#fixProfilePoint","label":"IDebugProfiler::fixProfilePoint()"},
 {"file":"interface-SmartFactory.Interfaces.IDebugProfiler.html#init","label":"IDebugProfiler::init()"},
-{"file":"interface-SmartFactory.Interfaces.IDebugProfiler.html#logMessageToFile","label":"IDebugProfiler::logMessageToFile()"},
 {"file":"interface-SmartFactory.Interfaces.IDebugProfiler.html#startProfilePoint","label":"IDebugProfiler::startProfilePoint()"},
 {"file":"interface-SmartFactory.Interfaces.IErrorHandler.html","label":"IErrorHandler - Interface"},
 {"file":"interface-SmartFactory.Interfaces.IErrorHandler.html#disableTrace","label":"IErrorHandler::disableTrace()"},
@@ -250,82 +320,79 @@ ApiGen.elements = [
 {"file":"interface-SmartFactory.Interfaces.IInitable.html","label":"IInitable - Interface"},
 {"file":"interface-SmartFactory.Interfaces.IInitable.html#init","label":"IInitable::init()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html","label":"ILanguageManager - Interface"},
+{"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#addLocalizationFile","label":"ILanguageManager::addLocalizationFile()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#detectLanguage","label":"ILanguageManager::detectLanguage()"},
+{"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#extendDictionary","label":"ILanguageManager::extendDictionary()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#getContext","label":"ILanguageManager::getContext()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#getCountryCode","label":"ILanguageManager::getCountryCode()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#getCountryList","label":"ILanguageManager::getCountryList()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#getCountryName","label":"ILanguageManager::getCountryName()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#getCurrentLanguage","label":"ILanguageManager::getCurrentLanguage()"},
+{"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#getFallbackLanguage","label":"ILanguageManager::getFallbackLanguage()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#getLanguageCode","label":"ILanguageManager::getLanguageCode()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#getLanguageList","label":"ILanguageManager::getLanguageList()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#getLanguageName","label":"ILanguageManager::getLanguageName()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#getSupportedLanguages","label":"ILanguageManager::getSupportedLanguages()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#hasTranslation","label":"ILanguageManager::hasTranslation()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#init","label":"ILanguageManager::init()"},
+{"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#loadDictionary","label":"ILanguageManager::loadDictionary()"},
+{"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#setContext","label":"ILanguageManager::setContext()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#setCurrentLanguage","label":"ILanguageManager::setCurrentLanguage()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#text","label":"ILanguageManager::text()"},
+{"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#try_text","label":"ILanguageManager::try_text()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#validateCountryCode","label":"ILanguageManager::validateCountryCode()"},
 {"file":"interface-SmartFactory.Interfaces.ILanguageManager.html#validateLanguageCode","label":"ILanguageManager::validateLanguageCode()"},
 {"file":"interface-SmartFactory.Interfaces.IMessageManager.html","label":"IMessageManager - Interface"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#addMessagesToResponse","label":"IMessageManager::addMessagesToResponse()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#addBubbleMessage","label":"IMessageManager::addBubbleMessage()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#addDebugMessage","label":"IMessageManager::addDebugMessage()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#addError","label":"IMessageManager::addError()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#addErrorFromSmartException","label":"IMessageManager::addErrorFromSmartException()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#addInfoMessage","label":"IMessageManager::addInfoMessage()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#addProgWarning","label":"IMessageManager::addProgWarning()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#addWarning","label":"IMessageManager::addWarning()"},
 {"file":"interface-SmartFactory.Interfaces.IMessageManager.html#clearAll","label":"IMessageManager::clearAll()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#clearBubbleMessages","label":"IMessageManager::clearBubbleMessages()"},
 {"file":"interface-SmartFactory.Interfaces.IMessageManager.html#clearDebugMessages","label":"IMessageManager::clearDebugMessages()"},
 {"file":"interface-SmartFactory.Interfaces.IMessageManager.html#clearErrors","label":"IMessageManager::clearErrors()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#clearInfos","label":"IMessageManager::clearInfos()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#clearInfoMessages","label":"IMessageManager::clearInfoMessages()"},
 {"file":"interface-SmartFactory.Interfaces.IMessageManager.html#clearProgWarnings","label":"IMessageManager::clearProgWarnings()"},
 {"file":"interface-SmartFactory.Interfaces.IMessageManager.html#clearWarnings","label":"IMessageManager::clearWarnings()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#debugMessageExists","label":"IMessageManager::debugMessageExists()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#detailsActive","label":"IMessageManager::detailsActive()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#disableDetails","label":"IMessageManager::disableDetails()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#disableProgWarnings","label":"IMessageManager::disableProgWarnings()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#enableDetails","label":"IMessageManager::enableDetails()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#enableProgWarnings","label":"IMessageManager::enableProgWarnings()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#errorsExist","label":"IMessageManager::errorsExist()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#getActiveTab","label":"IMessageManager::getActiveTab()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#getAutoHideTime","label":"IMessageManager::getAutoHideTime()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#getBubbleMessages","label":"IMessageManager::getBubbleMessages()"},
 {"file":"interface-SmartFactory.Interfaces.IMessageManager.html#getDebugMessages","label":"IMessageManager::getDebugMessages()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#getErrorElement","label":"IMessageManager::getErrorElement()"},
 {"file":"interface-SmartFactory.Interfaces.IMessageManager.html#getErrors","label":"IMessageManager::getErrors()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#getFocusElement","label":"IMessageManager::getFocusElement()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#getInfos","label":"IMessageManager::getInfos()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#getInfoMessages","label":"IMessageManager::getInfoMessages()"},
 {"file":"interface-SmartFactory.Interfaces.IMessageManager.html#getProgWarnings","label":"IMessageManager::getProgWarnings()"},
 {"file":"interface-SmartFactory.Interfaces.IMessageManager.html#getWarnings","label":"IMessageManager::getWarnings()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#infosExist","label":"IMessageManager::infosExist()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#hasBubbleMessages","label":"IMessageManager::hasBubbleMessages()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#hasDebugMessages","label":"IMessageManager::hasDebugMessages()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#hasErrors","label":"IMessageManager::hasErrors()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#hasInfoMessages","label":"IMessageManager::hasInfoMessages()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#hasProgWarnings","label":"IMessageManager::hasProgWarnings()"},
+{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#hasWarnings","label":"IMessageManager::hasWarnings()"},
 {"file":"interface-SmartFactory.Interfaces.IMessageManager.html#init","label":"IMessageManager::init()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#progWarningsActive","label":"IMessageManager::progWarningsActive()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#progWarningsExist","label":"IMessageManager::progWarningsExist()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#setActiveTab","label":"IMessageManager::setActiveTab()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#setDebugMessage","label":"IMessageManager::setDebugMessage()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#setError","label":"IMessageManager::setError()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#setErrorElement","label":"IMessageManager::setErrorElement()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#setFocusElement","label":"IMessageManager::setFocusElement()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#setInfo","label":"IMessageManager::setInfo()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#setProgWarning","label":"IMessageManager::setProgWarning()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#setWarning","label":"IMessageManager::setWarning()"},
-{"file":"interface-SmartFactory.Interfaces.IMessageManager.html#warningsExist","label":"IMessageManager::warningsExist()"},
 {"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html","label":"IRecordsetManager - Interface"},
+{"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#commit_transaction","label":"IRecordsetManager::commit_transaction()"},
+{"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#countRecords","label":"IRecordsetManager::countRecords()"},
+{"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#countRecordsQuery","label":"IRecordsetManager::countRecordsQuery()"},
 {"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#deleteRecords","label":"IRecordsetManager::deleteRecords()"},
+{"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#deleteRecordsQuery","label":"IRecordsetManager::deleteRecordsQuery()"},
 {"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#describeTableFields","label":"IRecordsetManager::describeTableFields()"},
+{"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#describeTableFieldsQuery","label":"IRecordsetManager::describeTableFieldsQuery()"},
+{"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#escape","label":"IRecordsetManager::escape()"},
+{"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#format_date","label":"IRecordsetManager::format_date()"},
+{"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#format_datetime","label":"IRecordsetManager::format_datetime()"},
 {"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#getDBWorker","label":"IRecordsetManager::getDBWorker()"},
 {"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#loadRecord","label":"IRecordsetManager::loadRecord()"},
+{"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#loadRecordQuery","label":"IRecordsetManager::loadRecordQuery()"},
 {"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#loadRecordSet","label":"IRecordsetManager::loadRecordSet()"},
+{"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#loadRecordSetQuery","label":"IRecordsetManager::loadRecordSetQuery()"},
+{"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#rollback_transaction","label":"IRecordsetManager::rollback_transaction()"},
 {"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#saveRecord","label":"IRecordsetManager::saveRecord()"},
 {"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#saveRecordSet","label":"IRecordsetManager::saveRecordSet()"},
 {"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#setDBWorker","label":"IRecordsetManager::setDBWorker()"},
+{"file":"interface-SmartFactory.Interfaces.IRecordsetManager.html#start_transaction","label":"IRecordsetManager::start_transaction()"},
 {"file":"interface-SmartFactory.Interfaces.IRequestHandler.html","label":"IRequestHandler - Interface"},
 {"file":"interface-SmartFactory.Interfaces.IRequestHandler.html#handleRequest","label":"IRequestHandler::handleRequest()"},
-{"file":"interface-SmartFactory.Interfaces.ISessionManager.html","label":"ISessionManager - Interface"},
-{"file":"interface-SmartFactory.Interfaces.ISessionManager.html#clearSession","label":"ISessionManager::clearSession()"},
-{"file":"interface-SmartFactory.Interfaces.ISessionManager.html#destroySession","label":"ISessionManager::destroySession()"},
-{"file":"interface-SmartFactory.Interfaces.ISessionManager.html#getContext","label":"ISessionManager::getContext()"},
-{"file":"interface-SmartFactory.Interfaces.ISessionManager.html#getSessionId","label":"ISessionManager::getSessionId()"},
-{"file":"interface-SmartFactory.Interfaces.ISessionManager.html#getSessionName","label":"ISessionManager::getSessionName()"},
-{"file":"interface-SmartFactory.Interfaces.ISessionManager.html#setSessionId","label":"ISessionManager::setSessionId()"},
-{"file":"interface-SmartFactory.Interfaces.ISessionManager.html#setSessionName","label":"ISessionManager::setSessionName()"},
-{"file":"interface-SmartFactory.Interfaces.ISessionManager.html#startSession","label":"ISessionManager::startSession()"},
-{"file":"interface-SmartFactory.Interfaces.ISessionManager.html#switchContext","label":"ISessionManager::switchContext()"},
-{"file":"interface-SmartFactory.Interfaces.ISessionManager.html#vars","label":"ISessionManager::vars()"},
-{"file":"interface-SmartFactory.Interfaces.ISessionManager.html#writeCloseSession","label":"ISessionManager::writeCloseSession()"},
 {"file":"interface-SmartFactory.Interfaces.ISettingsManager.html","label":"ISettingsManager - Interface"},
 {"file":"interface-SmartFactory.Interfaces.ISettingsManager.html#getContext","label":"ISettingsManager::getContext()"},
 {"file":"interface-SmartFactory.Interfaces.ISettingsManager.html#getParameter","label":"ISettingsManager::getParameter()"},
@@ -345,24 +412,14 @@ ApiGen.elements = [
 {"file":"interface-SmartFactory.Interfaces.IShardManager.html#randomDBShard","label":"IShardManager::randomDBShard()"},
 {"file":"interface-SmartFactory.Interfaces.IShardManager.html#registerShard","label":"IShardManager::registerShard()"},
 {"file":"class-SmartFactory.JsonRequestHandler.html","label":"JsonRequestHandler - Class"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#$action","label":"JsonRequestHandler::$action"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#$errors","label":"JsonRequestHandler::$errors"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#$param_string","label":"JsonRequestHandler::$param_string"},
 {"file":"class-SmartFactory.JsonRequestHandler.html#$request_data","label":"JsonRequestHandler::$request_data"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#$request_headers","label":"JsonRequestHandler::$request_headers"},
 {"file":"class-SmartFactory.JsonRequestHandler.html#$response_data","label":"JsonRequestHandler::$response_data"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#$response_headers","label":"JsonRequestHandler::$response_headers"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#addError","label":"JsonRequestHandler::addError()"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#addExceptionError","label":"JsonRequestHandler::addExceptionError()"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#exitWithErrors","label":"JsonRequestHandler::exitWithErrors()"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#handleRequest","label":"JsonRequestHandler::handleRequest()"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#parseJsonInput","label":"JsonRequestHandler::parseJsonInput()"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#postprocessRequest","label":"JsonRequestHandler::postprocessRequest()"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#preprocessRequest","label":"JsonRequestHandler::preprocessRequest()"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#processInputData","label":"JsonRequestHandler::processInputData()"},
-{"file":"class-SmartFactory.JsonRequestHandler.html#sendJsonResponse","label":"JsonRequestHandler::sendJsonResponse()"},
+{"file":"class-SmartFactory.JsonRequestHandler.html#parseInput","label":"JsonRequestHandler::parseInput()"},
+{"file":"class-SmartFactory.JsonRequestHandler.html#sendResponse","label":"JsonRequestHandler::sendResponse()"},
 {"file":"class-SmartFactory.LanguageManager.html","label":"LanguageManager - Class"},
+{"file":"class-SmartFactory.LanguageManager.html#$additional_localization_files","label":"LanguageManager::$additional_localization_files"},
 {"file":"class-SmartFactory.LanguageManager.html#$context","label":"LanguageManager::$context"},
+{"file":"class-SmartFactory.LanguageManager.html#$cookie_path","label":"LanguageManager::$cookie_path"},
 {"file":"class-SmartFactory.LanguageManager.html#$countries","label":"LanguageManager::$countries"},
 {"file":"class-SmartFactory.LanguageManager.html#$current_language","label":"LanguageManager::$current_language"},
 {"file":"class-SmartFactory.LanguageManager.html#$dictionary_loaded","label":"LanguageManager::$dictionary_loaded"},
@@ -371,12 +428,18 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.LanguageManager.html#$supported_languages","label":"LanguageManager::$supported_languages"},
 {"file":"class-SmartFactory.LanguageManager.html#$texts","label":"LanguageManager::$texts"},
 {"file":"class-SmartFactory.LanguageManager.html#$use_apcu","label":"LanguageManager::$use_apcu"},
+{"file":"class-SmartFactory.LanguageManager.html#$use_cookie","label":"LanguageManager::$use_cookie"},
+{"file":"class-SmartFactory.LanguageManager.html#$use_fallback_language","label":"LanguageManager::$use_fallback_language"},
+{"file":"class-SmartFactory.LanguageManager.html#$warn_missing","label":"LanguageManager::$warn_missing"},
+{"file":"class-SmartFactory.LanguageManager.html#addLocalizationFile","label":"LanguageManager::addLocalizationFile()"},
 {"file":"class-SmartFactory.LanguageManager.html#detectLanguage","label":"LanguageManager::detectLanguage()"},
+{"file":"class-SmartFactory.LanguageManager.html#extendDictionary","label":"LanguageManager::extendDictionary()"},
 {"file":"class-SmartFactory.LanguageManager.html#getContext","label":"LanguageManager::getContext()"},
 {"file":"class-SmartFactory.LanguageManager.html#getCountryCode","label":"LanguageManager::getCountryCode()"},
 {"file":"class-SmartFactory.LanguageManager.html#getCountryList","label":"LanguageManager::getCountryList()"},
 {"file":"class-SmartFactory.LanguageManager.html#getCountryName","label":"LanguageManager::getCountryName()"},
 {"file":"class-SmartFactory.LanguageManager.html#getCurrentLanguage","label":"LanguageManager::getCurrentLanguage()"},
+{"file":"class-SmartFactory.LanguageManager.html#getFallbackLanguage","label":"LanguageManager::getFallbackLanguage()"},
 {"file":"class-SmartFactory.LanguageManager.html#getLanguageCode","label":"LanguageManager::getLanguageCode()"},
 {"file":"class-SmartFactory.LanguageManager.html#getLanguageList","label":"LanguageManager::getLanguageList()"},
 {"file":"class-SmartFactory.LanguageManager.html#getLanguageName","label":"LanguageManager::getLanguageName()"},
@@ -384,57 +447,50 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.LanguageManager.html#hasTranslation","label":"LanguageManager::hasTranslation()"},
 {"file":"class-SmartFactory.LanguageManager.html#init","label":"LanguageManager::init()"},
 {"file":"class-SmartFactory.LanguageManager.html#loadDictionary","label":"LanguageManager::loadDictionary()"},
+{"file":"class-SmartFactory.LanguageManager.html#setContext","label":"LanguageManager::setContext()"},
 {"file":"class-SmartFactory.LanguageManager.html#setCurrentLanguage","label":"LanguageManager::setCurrentLanguage()"},
 {"file":"class-SmartFactory.LanguageManager.html#text","label":"LanguageManager::text()"},
+{"file":"class-SmartFactory.LanguageManager.html#try_text","label":"LanguageManager::try_text()"},
 {"file":"class-SmartFactory.LanguageManager.html#validateCountryCode","label":"LanguageManager::validateCountryCode()"},
 {"file":"class-SmartFactory.LanguageManager.html#validateLanguageCode","label":"LanguageManager::validateLanguageCode()"},
+{"file":"constant-SmartFactory.MAX_SIZE.html","label":"MAX_SIZE - Constant"},
 {"file":"class-SmartFactory.MessageManager.html","label":"MessageManager - Class"},
-{"file":"class-SmartFactory.MessageManager.html#$auto_hide_time","label":"MessageManager::$auto_hide_time"},
-{"file":"class-SmartFactory.MessageManager.html#$details_disabled","label":"MessageManager::$details_disabled"},
-{"file":"class-SmartFactory.MessageManager.html#$prog_warnings_disabled","label":"MessageManager::$prog_warnings_disabled"},
-{"file":"class-SmartFactory.MessageManager.html#$session_vars","label":"MessageManager::$session_vars"},
-{"file":"class-SmartFactory.MessageManager.html#__construct","label":"MessageManager::__construct()"},
-{"file":"class-SmartFactory.MessageManager.html#addMessagesToResponse","label":"MessageManager::addMessagesToResponse()"},
+{"file":"class-SmartFactory.MessageManager.html#$debug_mode","label":"MessageManager::$debug_mode"},
+{"file":"class-SmartFactory.MessageManager.html#$messages","label":"MessageManager::$messages"},
+{"file":"class-SmartFactory.MessageManager.html#$write_source_file_and_line_by_debug","label":"MessageManager::$write_source_file_and_line_by_debug"},
+{"file":"class-SmartFactory.MessageManager.html#addBubbleMessage","label":"MessageManager::addBubbleMessage()"},
+{"file":"class-SmartFactory.MessageManager.html#addDebugMessage","label":"MessageManager::addDebugMessage()"},
+{"file":"class-SmartFactory.MessageManager.html#addError","label":"MessageManager::addError()"},
+{"file":"class-SmartFactory.MessageManager.html#addErrorFromSmartException","label":"MessageManager::addErrorFromSmartException()"},
+{"file":"class-SmartFactory.MessageManager.html#addInfoMessage","label":"MessageManager::addInfoMessage()"},
+{"file":"class-SmartFactory.MessageManager.html#addProgWarning","label":"MessageManager::addProgWarning()"},
+{"file":"class-SmartFactory.MessageManager.html#addWarning","label":"MessageManager::addWarning()"},
 {"file":"class-SmartFactory.MessageManager.html#clearAll","label":"MessageManager::clearAll()"},
+{"file":"class-SmartFactory.MessageManager.html#clearBubbleMessages","label":"MessageManager::clearBubbleMessages()"},
 {"file":"class-SmartFactory.MessageManager.html#clearDebugMessages","label":"MessageManager::clearDebugMessages()"},
 {"file":"class-SmartFactory.MessageManager.html#clearErrors","label":"MessageManager::clearErrors()"},
-{"file":"class-SmartFactory.MessageManager.html#clearInfos","label":"MessageManager::clearInfos()"},
+{"file":"class-SmartFactory.MessageManager.html#clearInfoMessages","label":"MessageManager::clearInfoMessages()"},
 {"file":"class-SmartFactory.MessageManager.html#clearProgWarnings","label":"MessageManager::clearProgWarnings()"},
 {"file":"class-SmartFactory.MessageManager.html#clearWarnings","label":"MessageManager::clearWarnings()"},
-{"file":"class-SmartFactory.MessageManager.html#debugMessageExists","label":"MessageManager::debugMessageExists()"},
-{"file":"class-SmartFactory.MessageManager.html#detailsActive","label":"MessageManager::detailsActive()"},
-{"file":"class-SmartFactory.MessageManager.html#disableDetails","label":"MessageManager::disableDetails()"},
-{"file":"class-SmartFactory.MessageManager.html#disableProgWarnings","label":"MessageManager::disableProgWarnings()"},
-{"file":"class-SmartFactory.MessageManager.html#enableDetails","label":"MessageManager::enableDetails()"},
-{"file":"class-SmartFactory.MessageManager.html#enableProgWarnings","label":"MessageManager::enableProgWarnings()"},
-{"file":"class-SmartFactory.MessageManager.html#errorsExist","label":"MessageManager::errorsExist()"},
-{"file":"class-SmartFactory.MessageManager.html#extractMessagesForDisplay","label":"MessageManager::extractMessagesForDisplay()"},
-{"file":"class-SmartFactory.MessageManager.html#getActiveTab","label":"MessageManager::getActiveTab()"},
-{"file":"class-SmartFactory.MessageManager.html#getAutoHideTime","label":"MessageManager::getAutoHideTime()"},
+{"file":"class-SmartFactory.MessageManager.html#getBubbleMessages","label":"MessageManager::getBubbleMessages()"},
 {"file":"class-SmartFactory.MessageManager.html#getDebugMessages","label":"MessageManager::getDebugMessages()"},
-{"file":"class-SmartFactory.MessageManager.html#getErrorElement","label":"MessageManager::getErrorElement()"},
 {"file":"class-SmartFactory.MessageManager.html#getErrors","label":"MessageManager::getErrors()"},
-{"file":"class-SmartFactory.MessageManager.html#getFocusElement","label":"MessageManager::getFocusElement()"},
-{"file":"class-SmartFactory.MessageManager.html#getInfos","label":"MessageManager::getInfos()"},
+{"file":"class-SmartFactory.MessageManager.html#getInfoMessages","label":"MessageManager::getInfoMessages()"},
+{"file":"class-SmartFactory.MessageManager.html#getMessages","label":"MessageManager::getMessages()"},
 {"file":"class-SmartFactory.MessageManager.html#getProgWarnings","label":"MessageManager::getProgWarnings()"},
 {"file":"class-SmartFactory.MessageManager.html#getWarnings","label":"MessageManager::getWarnings()"},
-{"file":"class-SmartFactory.MessageManager.html#infosExist","label":"MessageManager::infosExist()"},
+{"file":"class-SmartFactory.MessageManager.html#hasBubbleMessages","label":"MessageManager::hasBubbleMessages()"},
+{"file":"class-SmartFactory.MessageManager.html#hasDebugMessages","label":"MessageManager::hasDebugMessages()"},
+{"file":"class-SmartFactory.MessageManager.html#hasErrors","label":"MessageManager::hasErrors()"},
+{"file":"class-SmartFactory.MessageManager.html#hasInfoMessages","label":"MessageManager::hasInfoMessages()"},
+{"file":"class-SmartFactory.MessageManager.html#hasProgWarnings","label":"MessageManager::hasProgWarnings()"},
+{"file":"class-SmartFactory.MessageManager.html#hasWarnings","label":"MessageManager::hasWarnings()"},
 {"file":"class-SmartFactory.MessageManager.html#init","label":"MessageManager::init()"},
-{"file":"class-SmartFactory.MessageManager.html#progWarningsActive","label":"MessageManager::progWarningsActive()"},
-{"file":"class-SmartFactory.MessageManager.html#progWarningsExist","label":"MessageManager::progWarningsExist()"},
-{"file":"class-SmartFactory.MessageManager.html#setActiveTab","label":"MessageManager::setActiveTab()"},
-{"file":"class-SmartFactory.MessageManager.html#setDebugMessage","label":"MessageManager::setDebugMessage()"},
-{"file":"class-SmartFactory.MessageManager.html#setError","label":"MessageManager::setError()"},
-{"file":"class-SmartFactory.MessageManager.html#setErrorElement","label":"MessageManager::setErrorElement()"},
-{"file":"class-SmartFactory.MessageManager.html#setFocusElement","label":"MessageManager::setFocusElement()"},
-{"file":"class-SmartFactory.MessageManager.html#setInfo","label":"MessageManager::setInfo()"},
-{"file":"class-SmartFactory.MessageManager.html#setProgWarning","label":"MessageManager::setProgWarning()"},
-{"file":"class-SmartFactory.MessageManager.html#setWarning","label":"MessageManager::setWarning()"},
-{"file":"class-SmartFactory.MessageManager.html#warningsExist","label":"MessageManager::warningsExist()"},
 {"file":"class-SmartFactory.ObjectFactory.html","label":"ObjectFactory - Class"},
 {"file":"class-SmartFactory.ObjectFactory.html#$itable","label":"ObjectFactory::$itable"},
 {"file":"class-SmartFactory.ObjectFactory.html#$singletons","label":"ObjectFactory::$singletons"},
 {"file":"class-SmartFactory.ObjectFactory.html#bindClass","label":"ObjectFactory::bindClass()"},
+{"file":"class-SmartFactory.ObjectFactory.html#bindClassContext","label":"ObjectFactory::bindClassContext()"},
 {"file":"class-SmartFactory.ObjectFactory.html#getInstance","label":"ObjectFactory::getInstance()"},
 {"file":"class-SmartFactory.RecordsetManager.html","label":"RecordsetManager - Class"},
 {"file":"class-SmartFactory.RecordsetManager.html#$dbworker","label":"RecordsetManager::$dbworker"},
@@ -442,16 +498,42 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.RecordsetManager.html#$key_fields","label":"RecordsetManager::$key_fields"},
 {"file":"class-SmartFactory.RecordsetManager.html#$table","label":"RecordsetManager::$table"},
 {"file":"class-SmartFactory.RecordsetManager.html#checkWhereClause","label":"RecordsetManager::checkWhereClause()"},
+{"file":"class-SmartFactory.RecordsetManager.html#commit_transaction","label":"RecordsetManager::commit_transaction()"},
+{"file":"class-SmartFactory.RecordsetManager.html#countRecords","label":"RecordsetManager::countRecords()"},
+{"file":"class-SmartFactory.RecordsetManager.html#countRecordsQuery","label":"RecordsetManager::countRecordsQuery()"},
 {"file":"class-SmartFactory.RecordsetManager.html#deleteRecords","label":"RecordsetManager::deleteRecords()"},
+{"file":"class-SmartFactory.RecordsetManager.html#deleteRecordsQuery","label":"RecordsetManager::deleteRecordsQuery()"},
 {"file":"class-SmartFactory.RecordsetManager.html#describeTableFields","label":"RecordsetManager::describeTableFields()"},
+{"file":"class-SmartFactory.RecordsetManager.html#describeTableFieldsQuery","label":"RecordsetManager::describeTableFieldsQuery()"},
+{"file":"class-SmartFactory.RecordsetManager.html#escape","label":"RecordsetManager::escape()"},
+{"file":"class-SmartFactory.RecordsetManager.html#format_date","label":"RecordsetManager::format_date()"},
+{"file":"class-SmartFactory.RecordsetManager.html#format_datetime","label":"RecordsetManager::format_datetime()"},
 {"file":"class-SmartFactory.RecordsetManager.html#getDBWorker","label":"RecordsetManager::getDBWorker()"},
 {"file":"class-SmartFactory.RecordsetManager.html#loadRecord","label":"RecordsetManager::loadRecord()"},
+{"file":"class-SmartFactory.RecordsetManager.html#loadRecordQuery","label":"RecordsetManager::loadRecordQuery()"},
 {"file":"class-SmartFactory.RecordsetManager.html#loadRecordSet","label":"RecordsetManager::loadRecordSet()"},
+{"file":"class-SmartFactory.RecordsetManager.html#loadRecordSetQuery","label":"RecordsetManager::loadRecordSetQuery()"},
+{"file":"class-SmartFactory.RecordsetManager.html#number_or_null","label":"RecordsetManager::number_or_null()"},
 {"file":"class-SmartFactory.RecordsetManager.html#processSubarray","label":"RecordsetManager::processSubarray()"},
+{"file":"class-SmartFactory.RecordsetManager.html#quotes_or_null","label":"RecordsetManager::quotes_or_null()"},
+{"file":"class-SmartFactory.RecordsetManager.html#rollback_transaction","label":"RecordsetManager::rollback_transaction()"},
 {"file":"class-SmartFactory.RecordsetManager.html#saveRecord","label":"RecordsetManager::saveRecord()"},
 {"file":"class-SmartFactory.RecordsetManager.html#saveRecordSet","label":"RecordsetManager::saveRecordSet()"},
 {"file":"class-SmartFactory.RecordsetManager.html#setDBWorker","label":"RecordsetManager::setDBWorker()"},
+{"file":"class-SmartFactory.RecordsetManager.html#start_transaction","label":"RecordsetManager::start_transaction()"},
 {"file":"class-SmartFactory.RecordsetManager.html#validateParameters","label":"RecordsetManager::validateParameters()"},
+{"file":"class-SmartFactory.RequestHandler.html","label":"RequestHandler - Class"},
+{"file":"class-SmartFactory.RequestHandler.html#$action","label":"RequestHandler::$action"},
+{"file":"class-SmartFactory.RequestHandler.html#$param_string","label":"RequestHandler::$param_string"},
+{"file":"class-SmartFactory.RequestHandler.html#$response_headers","label":"RequestHandler::$response_headers"},
+{"file":"class-SmartFactory.RequestHandler.html#addMessagesToResponse","label":"RequestHandler::addMessagesToResponse()"},
+{"file":"class-SmartFactory.RequestHandler.html#addResponseHeader","label":"RequestHandler::addResponseHeader()"},
+{"file":"class-SmartFactory.RequestHandler.html#handleRequest","label":"RequestHandler::handleRequest()"},
+{"file":"class-SmartFactory.RequestHandler.html#parseInput","label":"RequestHandler::parseInput()"},
+{"file":"class-SmartFactory.RequestHandler.html#postprocessRequest","label":"RequestHandler::postprocessRequest()"},
+{"file":"class-SmartFactory.RequestHandler.html#preprocessRequest","label":"RequestHandler::preprocessRequest()"},
+{"file":"class-SmartFactory.RequestHandler.html#processInputData","label":"RequestHandler::processInputData()"},
+{"file":"class-SmartFactory.RequestHandler.html#sendResponse","label":"RequestHandler::sendResponse()"},
 {"file":"class-SmartFactory.RuntimeSettingsManager.html","label":"RuntimeSettingsManager - Class"},
 {"file":"class-SmartFactory.RuntimeSettingsManager.html#$context","label":"RuntimeSettingsManager::$context"},
 {"file":"class-SmartFactory.RuntimeSettingsManager.html#$dbworker","label":"RuntimeSettingsManager::$dbworker"},
@@ -473,50 +555,40 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.RuntimeSettingsManager.html#setValidator","label":"RuntimeSettingsManager::setValidator()"},
 {"file":"class-SmartFactory.RuntimeSettingsManager.html#validateParameters","label":"RuntimeSettingsManager::validateParameters()"},
 {"file":"class-SmartFactory.RuntimeSettingsManager.html#validateSettings","label":"RuntimeSettingsManager::validateSettings()"},
-{"file":"class-SmartFactory.SessionManager.html","label":"SessionManager - Class"},
-{"file":"class-SmartFactory.SessionManager.html#$context","label":"SessionManager::$context"},
-{"file":"class-SmartFactory.SessionManager.html#$readonly","label":"SessionManager::$readonly"},
-{"file":"class-SmartFactory.SessionManager.html#clearSession","label":"SessionManager::clearSession()"},
-{"file":"class-SmartFactory.SessionManager.html#close","label":"SessionManager::close()"},
-{"file":"class-SmartFactory.SessionManager.html#create_sid","label":"SessionManager::create_sid()"},
-{"file":"class-SmartFactory.SessionManager.html#destroy","label":"SessionManager::destroy()"},
-{"file":"class-SmartFactory.SessionManager.html#destroySession","label":"SessionManager::destroySession()"},
-{"file":"class-SmartFactory.SessionManager.html#gc","label":"SessionManager::gc()"},
-{"file":"class-SmartFactory.SessionManager.html#getContext","label":"SessionManager::getContext()"},
-{"file":"class-SmartFactory.SessionManager.html#getSessionId","label":"SessionManager::getSessionId()"},
-{"file":"class-SmartFactory.SessionManager.html#getSessionName","label":"SessionManager::getSessionName()"},
-{"file":"class-SmartFactory.SessionManager.html#open","label":"SessionManager::open()"},
-{"file":"class-SmartFactory.SessionManager.html#read","label":"SessionManager::read()"},
-{"file":"class-SmartFactory.SessionManager.html#setSessionId","label":"SessionManager::setSessionId()"},
-{"file":"class-SmartFactory.SessionManager.html#setSessionName","label":"SessionManager::setSessionName()"},
-{"file":"class-SmartFactory.SessionManager.html#startSession","label":"SessionManager::startSession()"},
-{"file":"class-SmartFactory.SessionManager.html#switchContext","label":"SessionManager::switchContext()"},
-{"file":"class-SmartFactory.SessionManager.html#vars","label":"SessionManager::vars()"},
-{"file":"class-SmartFactory.SessionManager.html#write","label":"SessionManager::write()"},
-{"file":"class-SmartFactory.SessionManager.html#writeCloseSession","label":"SessionManager::writeCloseSession()"},
 {"file":"class-SmartFactory.SmartException.html","label":"SmartException - Class"},
 {"file":"class-SmartFactory.SmartException.html#$error_code","label":"SmartException::$error_code"},
-{"file":"class-SmartFactory.SmartException.html#$error_type","label":"SmartException::$error_type"},
+{"file":"class-SmartFactory.SmartException.html#$error_details","label":"SmartException::$error_details"},
+{"file":"class-SmartFactory.SmartException.html#$error_element","label":"SmartException::$error_element"},
+{"file":"class-SmartFactory.SmartException.html#$technical_info","label":"SmartException::$technical_info"},
 {"file":"class-SmartFactory.SmartException.html#ERR_CODE_CONFIG_ERROR","label":"SmartException::ERR_CODE_CONFIG_ERROR - Constant"},
+{"file":"class-SmartFactory.SmartException.html#ERR_CODE_FILE_WRITE_ERROR","label":"SmartException::ERR_CODE_FILE_WRITE_ERROR - Constant"},
 {"file":"class-SmartFactory.SmartException.html#ERR_CODE_INVALID_CONTENT_TYPE","label":"SmartException::ERR_CODE_INVALID_CONTENT_TYPE - Constant"},
 {"file":"class-SmartFactory.SmartException.html#ERR_CODE_INVALID_REQUEST_DATA","label":"SmartException::ERR_CODE_INVALID_REQUEST_DATA - Constant"},
 {"file":"class-SmartFactory.SmartException.html#ERR_CODE_JSON_PARSE_ERROR","label":"SmartException::ERR_CODE_JSON_PARSE_ERROR - Constant"},
 {"file":"class-SmartFactory.SmartException.html#ERR_CODE_MISSING_REQUEST_DATA","label":"SmartException::ERR_CODE_MISSING_REQUEST_DATA - Constant"},
 {"file":"class-SmartFactory.SmartException.html#ERR_CODE_NOT_FOUND","label":"SmartException::ERR_CODE_NOT_FOUND - Constant"},
+{"file":"class-SmartFactory.SmartException.html#ERR_CODE_PATH_ACCESS_ERROR","label":"SmartException::ERR_CODE_PATH_ACCESS_ERROR - Constant"},
 {"file":"class-SmartFactory.SmartException.html#ERR_CODE_SYSTEM_ERROR","label":"SmartException::ERR_CODE_SYSTEM_ERROR - Constant"},
 {"file":"class-SmartFactory.SmartException.html#ERR_CODE_XML_PARSE_ERROR","label":"SmartException::ERR_CODE_XML_PARSE_ERROR - Constant"},
-{"file":"class-SmartFactory.SmartException.html#ERR_TYPE_PROGRAMMING_ERROR","label":"SmartException::ERR_TYPE_PROGRAMMING_ERROR - Constant"},
-{"file":"class-SmartFactory.SmartException.html#ERR_TYPE_USER_ERROR","label":"SmartException::ERR_TYPE_USER_ERROR - Constant"},
 {"file":"class-SmartFactory.SmartException.html#__construct","label":"SmartException::__construct()"},
 {"file":"class-SmartFactory.SmartException.html#getErrorCode","label":"SmartException::getErrorCode()"},
-{"file":"class-SmartFactory.SmartException.html#getErrorType","label":"SmartException::getErrorType()"},
+{"file":"class-SmartFactory.SmartException.html#getErrorDetails","label":"SmartException::getErrorDetails()"},
+{"file":"class-SmartFactory.SmartException.html#getErrorElement","label":"SmartException::getErrorElement()"},
+{"file":"class-SmartFactory.SmartException.html#getTechnicalInfo","label":"SmartException::getTechnicalInfo()"},
+{"file":"class-SmartFactory.SmartException.html#setErrorElement","label":"SmartException::setErrorElement()"},
+{"file":"class-SmartFactory.SmartExceptionCollection.html","label":"SmartExceptionCollection - Class"},
+{"file":"class-SmartFactory.SmartExceptionCollection.html#$errors","label":"SmartExceptionCollection::$errors"},
+{"file":"class-SmartFactory.SmartExceptionCollection.html#addError","label":"SmartExceptionCollection::addError()"},
+{"file":"class-SmartFactory.SmartExceptionCollection.html#getErrors","label":"SmartExceptionCollection::getErrors()"},
+{"file":"class-SmartFactory.SmartExceptionCollection.html#throwIfErrors","label":"SmartExceptionCollection::throwIfErrors()"},
 {"file":"trait-SmartFactory.TestTrait.html","label":"TestTrait - Trait"},
-{"file":"trait-SmartFactory.TestTrait.html#sayGoodbye","label":"TestTrait::sayGoodbye()"},
+{"file":"trait-SmartFactory.TestTrait.html#sayGoodbuy","label":"TestTrait::sayGoodbuy()"},
 {"file":"trait-SmartFactory.TestTrait.html#sayHello","label":"TestTrait::sayHello()"},
 {"file":"class-SmartFactory.UserSettingsManager.html","label":"UserSettingsManager - Class"},
 {"file":"class-SmartFactory.UserSettingsManager.html#$context","label":"UserSettingsManager::$context"},
 {"file":"class-SmartFactory.UserSettingsManager.html#$dbworker","label":"UserSettingsManager::$dbworker"},
 {"file":"class-SmartFactory.UserSettingsManager.html#$multichoice_tables","label":"UserSettingsManager::$multichoice_tables"},
+{"file":"class-SmartFactory.UserSettingsManager.html#$no_own_transaction","label":"UserSettingsManager::$no_own_transaction"},
 {"file":"class-SmartFactory.UserSettingsManager.html#$settings","label":"UserSettingsManager::$settings"},
 {"file":"class-SmartFactory.UserSettingsManager.html#$settings_tables","label":"UserSettingsManager::$settings_tables"},
 {"file":"class-SmartFactory.UserSettingsManager.html#$user_id","label":"UserSettingsManager::$user_id"},
@@ -531,6 +603,7 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.UserSettingsManager.html#saveSettings","label":"UserSettingsManager::saveSettings()"},
 {"file":"class-SmartFactory.UserSettingsManager.html#saveSettingsData","label":"UserSettingsManager::saveSettingsData()"},
 {"file":"class-SmartFactory.UserSettingsManager.html#setContext","label":"UserSettingsManager::setContext()"},
+{"file":"class-SmartFactory.UserSettingsManager.html#setDBWorker","label":"UserSettingsManager::setDBWorker()"},
 {"file":"class-SmartFactory.UserSettingsManager.html#setParameter","label":"UserSettingsManager::setParameter()"},
 {"file":"class-SmartFactory.UserSettingsManager.html#setParameters","label":"UserSettingsManager::setParameters()"},
 {"file":"class-SmartFactory.UserSettingsManager.html#setUserID","label":"UserSettingsManager::setUserID()"},
@@ -538,29 +611,16 @@ ApiGen.elements = [
 {"file":"class-SmartFactory.UserSettingsManager.html#validateParameters","label":"UserSettingsManager::validateParameters()"},
 {"file":"class-SmartFactory.UserSettingsManager.html#validateSettings","label":"UserSettingsManager::validateSettings()"},
 {"file":"class-SmartFactory.XmlRequestHandler.html","label":"XmlRequestHandler - Class"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#$action","label":"XmlRequestHandler::$action"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#$errors","label":"XmlRequestHandler::$errors"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#$param_string","label":"XmlRequestHandler::$param_string"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#$request_headers","label":"XmlRequestHandler::$request_headers"},
 {"file":"class-SmartFactory.XmlRequestHandler.html#$request_xmldoc","label":"XmlRequestHandler::$request_xmldoc"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#$response_headers","label":"XmlRequestHandler::$response_headers"},
 {"file":"class-SmartFactory.XmlRequestHandler.html#$response_xmldoc","label":"XmlRequestHandler::$response_xmldoc"},
 {"file":"class-SmartFactory.XmlRequestHandler.html#__construct","label":"XmlRequestHandler::__construct()"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#addError","label":"XmlRequestHandler::addError()"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#addExceptionError","label":"XmlRequestHandler::addExceptionError()"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#exitWithErrors","label":"XmlRequestHandler::exitWithErrors()"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#handleRequest","label":"XmlRequestHandler::handleRequest()"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#parseXmlInput","label":"XmlRequestHandler::parseXmlInput()"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#postprocessRequest","label":"XmlRequestHandler::postprocessRequest()"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#preprocessRequest","label":"XmlRequestHandler::preprocessRequest()"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#processInputData","label":"XmlRequestHandler::processInputData()"},
-{"file":"class-SmartFactory.XmlRequestHandler.html#sendXmlResponse","label":"XmlRequestHandler::sendXmlResponse()"},
+{"file":"class-SmartFactory.XmlRequestHandler.html#parseInput","label":"XmlRequestHandler::parseInput()"},
+{"file":"class-SmartFactory.XmlRequestHandler.html#sendResponse","label":"XmlRequestHandler::sendResponse()"},
 {"file":"function-SmartFactory.aes_256_decrypt.html","label":"aes_256_decrypt()"},
 {"file":"function-SmartFactory.aes_256_encrypt.html","label":"aes_256_encrypt()"},
 {"file":"function-SmartFactory.array_to_dom.html","label":"array_to_dom()"},
 {"file":"function-SmartFactory.array_to_json.html","label":"array_to_json()"},
 {"file":"function-SmartFactory.checkbox.html","label":"checkbox()"},
-{"file":"function-SmartFactory.checkempty.html","label":"checkempty()"},
 {"file":"function-SmartFactory.common_prefix.html","label":"common_prefix()"},
 {"file":"function-SmartFactory.config_settings.html","label":"config_settings()"},
 {"file":"function-SmartFactory.dbshard.html","label":"dbshard()"},
@@ -568,14 +628,13 @@ ApiGen.elements = [
 {"file":"function-SmartFactory.debug_message.html","label":"debug_message()"},
 {"file":"function-SmartFactory.debugger.html","label":"debugger()"},
 {"file":"function-SmartFactory.dom_to_array.html","label":"dom_to_array()"},
-{"file":"function-SmartFactory.echo_html.html","label":"echo_html()"},
-{"file":"function-SmartFactory.echo_js.html","label":"echo_js()"},
 {"file":"function-SmartFactory.escape_html.html","label":"escape_html()"},
 {"file":"function-SmartFactory.escape_html_array.html","label":"escape_html_array()"},
 {"file":"function-SmartFactory.escape_js.html","label":"escape_js()"},
 {"file":"function-SmartFactory.event.html","label":"event()"},
 {"file":"function-SmartFactory.format_number.html","label":"format_number()"},
 {"file":"function-SmartFactory.get_cookie.html","label":"get_cookie()"},
+{"file":"function-SmartFactory.get_header.html","label":"get_header()"},
 {"file":"function-SmartFactory.input_hidden.html","label":"input_hidden()"},
 {"file":"function-SmartFactory.input_password.html","label":"input_password()"},
 {"file":"function-SmartFactory.input_text.html","label":"input_text()"},
@@ -591,12 +650,15 @@ ApiGen.elements = [
 {"file":"function-SmartFactory.reqvar_value.html","label":"reqvar_value()"},
 {"file":"function-SmartFactory.runtime_settings.html","label":"runtime_settings()"},
 {"file":"function-SmartFactory.select.html","label":"select()"},
-{"file":"function-SmartFactory.session.html","label":"session()"},
 {"file":"function-SmartFactory.set_cookie.html","label":"set_cookie()"},
 {"file":"function-SmartFactory.singleton.html","label":"singleton()"},
+{"file":"function-SmartFactory.string_empty.html","label":"string_empty()"},
+{"file":"function-SmartFactory.string_to_number.html","label":"string_to_number()"},
 {"file":"function-SmartFactory.table.html","label":"table()"},
 {"file":"function-SmartFactory.text.html","label":"text()"},
 {"file":"function-SmartFactory.textarea.html","label":"textarea()"},
 {"file":"function-SmartFactory.timestamp.html","label":"timestamp()"},
-{"file":"function-SmartFactory.user_settings.html","label":"user_settings()"}
+{"file":"function-SmartFactory.user_settings.html","label":"user_settings()"},
+{"file":"function-preg_pattern_quote.html","label":"preg_pattern_quote()"},
+{"file":"function-preg_replacement_quote.html","label":"preg_replacement_quote()"}
 ];
