@@ -22,31 +22,14 @@ interface IDebugProfiler extends IInitable
      * @param array $parameters
      * The parameters may vary for each debug profiler.
      *
-     * @return boolean
-     * The method should return true upon successful initialization, otherwise false.
-     *
-     * @author Oleg Schildt
-     */
-    public function init($parameters);
-    
-    /**
-     * Logs a message to a specified log file.
-     *
-     * @param string $message
-     * Message to be logged.
-     *
-     * @param string $file_name
-     * The target file name.
-     *
-     * @return boolean
-     * It should return true if the logging was successful, otherwise false.
+     * @return void
      *
      * @throws \Exception
-     * It might throw an exception in the case of any errors.
+     * It might throw an exception in the case of any system errors.
      *
      * @author Oleg Schildt
      */
-    public function logMessageToFile($message, $file_name);
+    public function init(array $parameters): void;
     
     /**
      * Logs a message to a standard debug output.
@@ -54,15 +37,20 @@ interface IDebugProfiler extends IInitable
      * @param string $message
      * Message to be logged.
      *
-     * @return boolean
-     * It should return true if the logging was successful, otherwise false.
+     * @param bool $write_call_stack
+     * Flag that defines where the call stack should be written.
+     *
+     * @param string $file_name
+     * The target file name.
+     *
+     * @return void
      *
      * @throws \Exception
      * It might throw an exception in the case of any errors.
      *
      * @author Oleg Schildt
      */
-    public function debugMessage($message);
+    public function debugMessage(string $message, bool $write_call_stack = false, string $file_name = "debug.log"): void;
     
     /**
      * Logs a message to a standard profiling output and stores the current time.
@@ -70,8 +58,7 @@ interface IDebugProfiler extends IInitable
      * @param string $message
      * Message to be logged.
      *
-     * @return boolean
-     * It should return true if the logging was successful, otherwise false.
+     * @return void
      *
      * @throws \Exception
      * It might throw an exception in the case of any errors.
@@ -80,7 +67,7 @@ interface IDebugProfiler extends IInitable
      *
      * @author Oleg Schildt
      */
-    public function startProfilePoint($message);
+    public function startProfilePoint(string $message): void;
     
     /**
      * Logs a message to a standard profiling output and shows
@@ -90,8 +77,7 @@ interface IDebugProfiler extends IInitable
      * @param string $message
      * Message to be logged.
      *
-     * @return boolean
-     * It should return true if the logging was successful, otherwise false.
+     * @return void
      *
      * @throws \Exception
      * It might throw an exception in the case of any errors.
@@ -100,7 +86,7 @@ interface IDebugProfiler extends IInitable
      *
      * @author Oleg Schildt
      */
-    public function fixProfilePoint($message);
+    public function fixProfilePoint(string $message): void;
     
     /**
      * Clears the specified log file.
@@ -108,20 +94,22 @@ interface IDebugProfiler extends IInitable
      * @param string $file_name
      * The target file name.
      *
-     * @return boolean
-     * It should return true if the file has been successfully deleted, otherwise false.
+     * @return void
+     *
+     * @see IDebugProfiler::clearLogFiles()
      *
      * @author Oleg Schildt
      */
-    public function clearLogFile($file_name);
+    public function clearLogFile(string $file_name): void;
     
     /**
      * Clears all log files.
      *
-     * @return boolean
-     * It should return true if the files have been successfully deleted, otherwise false.
+     * @return void
+     *
+     * @see IDebugProfiler::clearLogFile()
      *
      * @author Oleg Schildt
      */
-    public function clearLogFiles();
+    public function clearLogFiles(): void;
 } // IDebugProfiler
