@@ -11,9 +11,9 @@
 
 namespace SmartFactory;
 
-use SmartFactory\Interfaces\IErrorHandler;
-use SmartFactory\Interfaces\IEventManager;
-use SmartFactory\Interfaces\IMessageManager;
+use \SmartFactory\Interfaces\IErrorHandler;
+use \SmartFactory\Interfaces\IEventManager;
+use \SmartFactory\Interfaces\IMessageManager;
 
 //------------------------------------------------------------------------------
 set_error_handler(function ($errno, $errstr, $errfile, $errline) {
@@ -21,6 +21,6 @@ set_error_handler(function ($errno, $errstr, $errfile, $errline) {
 });
 
 singleton(IEventManager::class)->addHandler("php_error", function ($event, $params) {
-    singleton(IMessageManager::class)->setProgWarning($params["etype"] . ": " . str_replace("<br/>", "\n", trim($params["errstr"])) . "\n\n" . $params["errfile"] . "\n" . "line " . $params["errline"]);
+    singleton(IMessageManager::class)->addProgWarning($params["etype"] . ": " . str_replace("<br/>", "\n", trim($params["errstr"])), $params["errfile"], $params["errline"]);
 });
 //------------------------------------------------------------------------------
