@@ -9,30 +9,31 @@
 
 namespace SmartFactory;
 
-use SmartFactory\Interfaces\IMessageManager;
-use SmartFactory\Interfaces\ISessionManager;
-use SmartFactory\Interfaces\IErrorHandler;
-use SmartFactory\Interfaces\IEventManager;
-use SmartFactory\Interfaces\IShardManager;
+use \SmartFactory\Interfaces\IMessageManager;
+use \SmartFactory\Interfaces\ISessionManager;
+use \SmartFactory\Interfaces\IErrorHandler;
+use \SmartFactory\Interfaces\IEventManager;
+use \SmartFactory\Interfaces\IShardManager;
 
-use SmartFactory\DatabaseWorkers\MySQL_DBWorker;
-use SmartFactory\DatabaseWorkers\MSSQL_DBWorker;
-use SmartFactory\DatabaseWorkers\ShardManager;
+use \SmartFactory\DatabaseWorkers\MySQL_DBWorker;
+use \SmartFactory\DatabaseWorkers\PostgreSQL_DBWorker;
+use \SmartFactory\DatabaseWorkers\MSSQL_DBWorker;
+use \SmartFactory\DatabaseWorkers\ShardManager;
 
 //-------------------------------------------------------------------
 // Class binding
 //-------------------------------------------------------------------
 ObjectFactory::bindClass(IErrorHandler::class, ErrorHandler::class);
 //-------------------------------------------------------------------
-ObjectFactory::bindClass(ISessionManager::class, SessionManager::class);
-//-------------------------------------------------------------------
 ObjectFactory::bindClass(IEventManager::class, EventManager::class);
 //-------------------------------------------------------------------
 ObjectFactory::bindClass(IMessageManager::class, MessageManager::class, function ($instance) {
-    $instance->init(["auto_hide_time" => 3]);
+    $instance->init(["debug_mode" => 1]);
 });
 //-------------------------------------------------------------------
 ObjectFactory::bindClass(MySQL_DBWorker::class, MySQL_DBWorker::class);
+//-------------------------------------------------------------------
+ObjectFactory::bindClass(PostgreSQL_DBWorker::class, PostgreSQL_DBWorker::class);
 //-------------------------------------------------------------------
 ObjectFactory::bindClass(MSSQL_DBWorker::class, MSSQL_DBWorker::class);
 //-------------------------------------------------------------------
